Function esp_idf_sys::esp_console_get_hint

pub unsafe extern "C" fn esp_console_get_hint(
    buf: *const c_char,
    color: *mut c_int,
    bold: *mut c_int,
) -> *const c_char

@brief Callback which provides command hints for linenoise library

When using linenoise for line editing, hints support can be enabled as follows:

linenoiseSetHintsCallback((linenoiseHintsCallback*) &esp_console_get_hint);

The extra cast is needed because linenoiseHintsCallback is defined as returning a char* instead of const char*.

@param buf line typed by the user @param[out] color ANSI color code to be used when displaying the hint @param[out] bold set to 1 if hint has to be displayed in bold @return string containing the hint text. This string is persistent and should not be freed (i.e. linenoiseSetFreeHintsCallback should not be used).