Function esp_idf_sys::esp_deep_sleep

pub unsafe extern "C" fn esp_deep_sleep(time_in_us: u64)

@brief Enter deep-sleep mode

The device will automatically wake up after the deep-sleep time Upon waking up, the device calls deep sleep wake stub, and then proceeds to load application.

Call to this function is equivalent to a call to esp_deep_sleep_enable_timer_wakeup followed by a call to esp_deep_sleep_start.

esp_deep_sleep does not shut down WiFi, BT, and higher level protocol connections gracefully. Make sure relevant WiFi and BT stack functions are called to close any connections and deinitialize the peripherals. These include: - esp_bluedroid_disable - esp_bt_controller_disable - esp_wifi_stop

This function does not return.

@note The device will wake up immediately if the deep-sleep time is set to 0

@param time_in_us deep-sleep time, unit: microsecond