esp_idf_sys
Function esp_eap_client_set_certificate_and_key

pub unsafe extern "C" fn esp_eap_client_set_certificate_and_key(
    client_cert: *const c_uchar,
    client_cert_len: c_int,
    private_key: *const c_uchar,
    private_key_len: c_int,
    private_key_password: *const c_uchar,
    private_key_passwd_len: c_int,
) -> esp_err_t

@brief Set client certificate and private key for EAP authentication.

This function sets the client certificate and private key to be used during authentication. Optionally, a private key password can be provided for encrypted private keys.

@attention 1. The client certificate, private key, and private key password are provided as pointers to the respective data arrays. @attention 2. The client_cert, private_key, and private_key_password should be zero-terminated.

@param[in] client_cert Pointer to the client certificate data. @param[in] client_cert_len Length of the client certificate data. @param[in] private_key Pointer to the private key data. @param[in] private_key_len Length of the private key data (limited to 1~4096 bytes). @param[in] private_key_password Pointer to the private key password data (optional). @param[in] private_key_passwd_len Length of the private key password data (can be 0 for no password).

@return

ESP_OK: The certificate, private key, and password (if provided) were set successfully.