esp_idf_sys

Function esp_eap_client_set_domain_name

pub unsafe extern "C" fn esp_eap_client_set_domain_name(
    domain_name: *const c_char,
) -> esp_err_t

@brief Set the domain name for certificate validation

This function sets the expected domain name for validating the certificate’s subject name. If the provided domain name does not match the certificate’s subject name, validation will fail.

@attention 1. The domain_name should be a NULL-terminated string.

@param[in] domain_name The expected domain name. Pass NULL to clear the domain matching.

@return

ESP_OK: The domain match was set successfully.

ESP_ERR_INVALID_ARG: Invalid argument (length > 255).

ESP_ERR_NO_MEM: Memory allocation failure.

ESP_ERR_NOT_SUPPORTED: Feature not supported.