esp_idf_sys
Function esp_efuse_write_field_cnt

pub unsafe extern "C" fn esp_efuse_write_field_cnt(
    field: *mut *const esp_efuse_desc_t,
    cnt: usize,
) -> esp_err_t

@brief Writes a required count of bits as “1” to EFUSE field.

If there are no free bits in the field to set the required number of bits to “1”, ESP_ERR_EFUSE_CNT_IS_FULL error is returned, the field will not be partially recorded. After the function is completed, the writing registers are cleared. @param[in] field A pointer to the structure describing the fields of efuse. @param[in] cnt Required number of programmed as “1” bits.

@return

ESP_OK: The operation was successfully completed.

ESP_ERR_INVALID_ARG: Error in the passed arguments.

ESP_ERR_EFUSE_CNT_IS_FULL: Not all requested cnt bits is set.