Function esp_idf_sys::esp_err_to_name_r

pub unsafe extern "C" fn esp_err_to_name_r(
    code: esp_err_t, 
    buf: *mut c_char, 
    buflen: size_t
) -> *const c_char

@brief Returns string for esp_err_t and system error codes

This function finds the error code in a pre-generated lookup-table of esp_err_t errors and returns its string representation. If the error code is not found then it is attempted to be found among system errors.

The function is generated by the Python script tools/gen_esp_err_to_name.py which should be run each time an esp_err_t error is modified, created or removed from the IDF project.

@param code esp_err_t error code @param[out] buf buffer where the error message should be written @param buflen Size of buffer buf. At most buflen bytes are written into the buf buffer (including the terminating null byte). @return buf containing the string error message