esp_idf_sys
Function esp_lcd_panel_draw_bitmap

pub unsafe extern "C" fn esp_lcd_panel_draw_bitmap(
    panel: esp_lcd_panel_handle_t,
    x_start: c_int,
    y_start: c_int,
    x_end: c_int,
    y_end: c_int,
    color_data: *const c_void,
) -> esp_err_t

@brief Draw bitmap on LCD panel

@param[in] panel LCD panel handle, which is created by other factory API like esp_lcd_new_panel_st7789() @param[in] x_start Start pixel index in the target frame buffer, on x-axis (x_start is included) @param[in] y_start Start pixel index in the target frame buffer, on y-axis (y_start is included) @param[in] x_end End pixel index in the target frame buffer, on x-axis (x_end is not included) @param[in] y_end End pixel index in the target frame buffer, on y-axis (y_end is not included) @param[in] color_data RGB color data that will be dumped to the specific window range @return - ESP_OK on success