Function esp_idf_sys::esp_mesh_switch_channel

pub unsafe extern "C" fn esp_mesh_switch_channel(
    new_bssid: *const u8,
    csa_newchan: c_int,
    csa_count: c_int
) -> esp_err_t

@brief Cause the root device to add Channel Switch Announcement Element (CSA IE) to beacon - Set the new channel - Set how many beacons with CSA IE will be sent before changing a new channel - Enable the channel switch function

@attention This API is only called by the root.

@param[in] new_bssid the new router BSSID if the router changes @param[in] csa_newchan the new channel number to which the whole network is moving @param[in] csa_count channel switch period(beacon count), unit is based on beacon interval of its softAP, the default value is 15.

@return

ESP_OK