esp_idf_sys
Function esp_mqtt_set_config

pub unsafe extern "C" fn esp_mqtt_set_config(
    client: esp_mqtt_client_handle_t,
    config: *const esp_mqtt_client_config_t,
) -> esp_err_t

@brief Set configuration structure, typically used when updating the config (i.e. on “before_connect” event

Notes:

When calling this function make sure to have all the intendend configurations set, otherwise default values are set. @param client MQTT client handle

@param config MQTT configuration structure

@return ESP_ERR_NO_MEM if failed to allocate ESP_ERR_INVALID_ARG if conflicts on transport configuration. ESP_OK on success