esp_idf_sys

Function esp_netif_action_got_ip

pub unsafe extern "C" fn esp_netif_action_got_ip(
    esp_netif: *mut c_void,
    base: esp_event_base_t,
    event_id: i32,
    data: *mut c_void,
)

@brief Default building block for network interface action upon network got IP event

@note This API can be directly used as event handler

@param[in] esp_netif Handle to esp-netif instance @param base The base type of the event @param event_id The specific ID of the event @param data Optional data associated with the event