Function esp_idf_sys::esp_partition_find_first

pub unsafe extern "C" fn esp_partition_find_first(
    type_: esp_partition_type_t, 
    subtype: esp_partition_subtype_t, 
    label: *const c_char
) -> *const esp_partition_t

@brief Find first partition based on one or more parameters

@param type Partition type, one of esp_partition_type_t values or an 8-bit unsigned integer @param subtype Partition subtype, one of esp_partition_subtype_t values or an 8-bit unsigned integer To find all partitions of given type, use ESP_PARTITION_SUBTYPE_ANY. @param label (optional) Partition label. Set this value if looking for partition with a specific name. Pass NULL otherwise.

@return pointer to esp_partition_t structure, or NULL if no partition is found. This pointer is valid for the lifetime of the application.