Function esp_idf_sys::esp_set_watchpoint

pub unsafe extern "C" fn esp_set_watchpoint(
    no: c_int, 
    adr: *mut c_void, 
    size: c_int, 
    flags: c_int
) -> esp_err_t

@brief Set a watchpoint to break/panic when a certain memory range is accessed.

@param no Watchpoint number. On the ESP32, this can be 0 or 1. @param adr Base address to watch @param size Size of the region, starting at the base address, to watch. Must be one of 2^n, with n in [0..6]. @param flags One of ESP_WATCHPOINT_* flags

@return ESP_ERR_INVALID_ARG on invalid arg, ESP_OK otherwise

@warning The ESP32 watchpoint hardware watches a region of bytes by effectively masking away the lower n bits for a region with size 2^n. If adr does not have zero for these lower n bits, you may not be watching the region you intended.