esp_idf_sys

Function esp_task_wdt_add_user

pub unsafe extern "C" fn esp_task_wdt_add_user(
    user_name: *const c_char,
    user_handle_ret: *mut esp_task_wdt_user_handle_t,
) -> esp_err_t

@brief Subscribe a user to the Task Watchdog Timer (TWDT)

This function subscribes a user to the TWDT. A user of the TWDT is usually a function that needs to run periodically. Each subscribed user must periodically call esp_task_wdt_reset_user() to prevent the TWDT from elapsing its timeout period. Failure to do so will result in a TWDT timeout.

@param[in] user_name String to identify the user @param[out] user_handle_ret Handle of the user @return

ESP_OK: Successfully subscribed the user to the TWDT

Other: Failed to subscribe user