Function esp_idf_sys::esp_task_wdt_init

pub unsafe extern "C" fn esp_task_wdt_init(
    timeout: u32, 
    panic: bool
) -> esp_err_t

@brief Initialize the Task Watchdog Timer (TWDT)

This function configures and initializes the TWDT. If the TWDT is already initialized when this function is called, this function will update the TWDT’s timeout period and panic configurations instead. After initializing the TWDT, any task can elect to be watched by the TWDT by subscribing to it using esp_task_wdt_add().

@param[in] timeout Timeout period of TWDT in seconds @param[in] panic Flag that controls whether the panic handler will be executed when the TWDT times out

@return - ESP_OK: Initialization was successful - ESP_ERR_NO_MEM: Initialization failed due to lack of memory

@note esp_task_wdt_init() must only be called after the scheduler started