Function esp_idf_sys::esp_tls_init_global_ca_store

pub unsafe extern "C" fn esp_tls_init_global_ca_store() -> esp_err_t

@brief Create a global CA store, initially empty.

This function should be called if the application wants to use the same CA store for multiple connections. This function initialises the global CA store which can be then set by calling esp_tls_set_global_ca_store(). To be effective, this function must be called before any call to esp_tls_set_global_ca_store().

@return - ESP_OK if creating global CA store was successful. - ESP_ERR_NO_MEM if an error occured when allocating the mbedTLS resources.