esp_idf_sys
Function esp_transport_ssl_set_psk_key_hint

pub unsafe extern "C" fn esp_transport_ssl_set_psk_key_hint(
    t: esp_transport_handle_t,
    psk_hint_key: *const psk_hint_key_t,
)

@brief Set PSK key and hint for PSK server/client verification in esp-tls component. Important notes: - This function stores the pointer to data, rather than making a copy. So this data must remain valid until after the connection is cleaned up - ESP_TLS_PSK_VERIFICATION config option must be enabled in menuconfig - certificate verification takes priority so it must not be configured to enable PSK method.

@param t ssl transport @param[in] psk_hint_key psk key and hint structure defined in esp_tls.h