Function esp_idf_sys::esp_vfs_register_fd_with_local_fd

pub unsafe extern "C" fn esp_vfs_register_fd_with_local_fd(
    vfs_id: esp_vfs_id_t,
    local_fd: c_int,
    permanent: bool,
    fd: *mut c_int,
) -> esp_err_t

Special function for registering another file descriptor with given local_fd for a VFS registered by esp_vfs_register_with_id.

@param vfs_id VFS identificator returned by esp_vfs_register_with_id. @param local_fd The fd in the local vfs. Passing -1 will set the local fd as the (*fd) value. @param permanent Whether the fd should be treated as permannet (not removed after close()) @param fd The registered file descriptor will be written to this address.

@return ESP_OK if the registration is successful, ESP_ERR_NO_MEM if too many file descriptors are registered, ESP_ERR_INVALID_ARG if the arguments are incorrect.