Function esp_idf_sys::essl_get_packet

pub unsafe extern "C" fn essl_get_packet(
    handle: essl_handle_t, 
    out_data: *mut c_void, 
    size: size_t, 
    out_length: *mut size_t, 
    wait_ms: u32
) -> esp_err_t

Get a packet from ESSL slave.

@param handle Handle of a essl device. @param[out] out_data Data output address @param size The size of the output buffer, if the buffer is smaller than the size of data to receive from slave, the driver returns ESP_ERR_NOT_FINISHED @param[out] out_length Output of length the data actually received from slave. @param wait_ms Millisecond to wait before timeout, will not wait at all if set to 0-9.

@return - ESP_OK Success: All the data has been read from the slave. - ESP_ERR_INVALID_ARG: Invalid argument, The handle is not initialized or the other arguments are invalid. - ESP_ERR_NOT_FINISHED: Read was successful, but there is still data remaining. - ESP_ERR_NOT_FOUND: Slave is not ready to send data. - ESP_ERR_NOT_SUPPORTED: This API is not supported in this mode - One of the error codes from SDMMC/SPI host controller.