esp_idf_sys
Function gpio_new_etm_event

pub unsafe extern "C" fn gpio_new_etm_event(
    config: *const gpio_etm_event_config_t,
    ret_event: *mut esp_etm_event_handle_t,
    ...
) -> esp_err_t

@brief Create an ETM event object for the GPIO peripheral

@note The created ETM event object can be deleted later by calling esp_etm_del_event @note The newly created ETM event object is not bind to any GPIO, you need to call gpio_etm_event_bind_gpio to bind the wanted GPIO @note Every success call to this function will acquire a free GPIO ETM event channel

@param[in] config GPIO ETM event configuration @param[out] ret_event Returned ETM event handle @param[out] … Other returned ETM event handles if any (the order of the returned event handles is aligned with the array order in field edges in gpio_etm_event_config_t) @return - ESP_OK: Create ETM event successfully - ESP_ERR_INVALID_ARG: Create ETM event failed because of invalid argument - ESP_ERR_NO_MEM: Create ETM event failed because of out of memory - ESP_ERR_NOT_FOUND: Create ETM event failed because all events are used up and no more free one - ESP_FAIL: Create ETM event failed because of other reasons