esp_idf_sys
Function heap_caps_malloc

pub unsafe extern "C" fn heap_caps_malloc(
    size: usize,
    caps: u32,
) -> *mut c_void

@brief Allocate a chunk of memory which has the given capabilities

Equivalent semantics to libc malloc(), for capability-aware memory.

@param size Size, in bytes, of the amount of memory to allocate @param caps Bitwise OR of MALLOC_CAP_* flags indicating the type of memory to be returned

@return A pointer to the memory allocated on success, NULL on failure