esp_idf_sys

Function i2c_master_get_bus_handle

pub unsafe extern "C" fn i2c_master_get_bus_handle(
    port_num: i2c_port_num_t,
    ret_handle: *mut i2c_master_bus_handle_t,
) -> esp_err_t

@brief Retrieves the I2C master bus handle for a specified I2C port number.

This function retrieves the I2C master bus handle for the given I2C port number. Please make sure the handle has already been initialized, and this function would simply returns the existing handle. Note that the returned handle still can’t be used concurrently

@param port_num I2C port number for which the handle is to be retrieved. @param ret_handle Pointer to a variable where the retrieved handle will be stored. @return - ESP_OK: Success. The handle is retrieved successfully. - ESP_ERR_INVALID_ARG: Invalid argument, such as invalid port number - ESP_ERR_INVALID_STATE: Invalid state, such as the I2C port is not initialized.