esp_idf_sys

Function i2c_master_write

pub unsafe extern "C" fn i2c_master_write(
    cmd_handle: i2c_cmd_handle_t,
    data: *const u8,
    data_len: usize,
    ack_en: bool,
) -> esp_err_t

@brief Queue a “write (multiple) bytes” command to the commands list. This function shall only be called in I2C master mode. Call i2c_master_cmd_begin() to send all queued commands

@param cmd_handle I2C commands list @param data Bytes to send. This buffer shall remain valid until the transaction is finished. If the PSRAM is enabled and intr_flag is set to ESP_INTR_FLAG_IRAM, data should be allocated from internal RAM. @param data_len Length, in bytes, of the data buffer @param ack_en Enable ACK signal

@return - ESP_OK Success - ESP_ERR_INVALID_ARG Parameter error - ESP_ERR_NO_MEM The static buffer used to create cmd_handler is too small - ESP_FAIL No more memory left on the heap