esp_idf_sys

Function i2s_channel_reconfig_pdm_rx_clock

pub unsafe extern "C" fn i2s_channel_reconfig_pdm_rx_clock(
    handle: i2s_chan_handle_t,
    clk_cfg: *const i2s_pdm_rx_clk_config_t,
) -> esp_err_t

@brief Reconfigure the I2S clock for PDM RX mode @note Only allowed to be called when the channel state is READY, i.e., channel has been initialized, but not started this function won’t change the state. i2s_channel_disable should be called before calling this function if I2S has started. @note The input channel handle has to be initialized to PDM RX mode, i.e., i2s_channel_init_pdm_rx_mode has been called before reconfiguring

@param[in] handle I2S RX channel handler @param[in] clk_cfg PDM RX mode clock configuration, can be generated by I2S_PDM_RX_CLK_DEFAULT_CONFIG @return - ESP_OK Set clock successfully - ESP_ERR_INVALID_ARG NULL pointer, invalid configuration or not PDM mode - ESP_ERR_INVALID_STATE This channel is not initialized or not stopped