Function esp_idf_sys::ledc_set_duty_with_hpoint

pub unsafe extern "C" fn ledc_set_duty_with_hpoint(
    speed_mode: ledc_mode_t, 
    channel: ledc_channel_t, 
    duty: u32, 
    hpoint: u32
) -> esp_err_t

@brief LEDC set duty and hpoint value Only after calling ledc_update_duty will the duty update. @note ledc_set_duty, ledc_set_duty_with_hpoint and ledc_update_duty are not thread-safe, do not call these functions to control one LEDC channel in different tasks at the same time. A thread-safe version of API is ledc_set_duty_and_update @note If a fade operation is running in progress on that channel, the driver would not allow it to be stopped. Other duty operations will have to wait until the fade operation has finished. @param speed_mode Select the LEDC channel group with specified speed mode. Note that not all targets support high speed mode. @param channel LEDC channel (0-7), select from ledc_channel_t @param duty Set the LEDC duty, the range of duty setting is [0, (2**duty_resolution)] @param hpoint Set the LEDC hpoint value(max: 0xfffff)

@return - ESP_OK Success - ESP_ERR_INVALID_ARG Parameter error