Function esp_idf_sys::lldesc_setup_link_constrained

pub unsafe extern "C" fn lldesc_setup_link_constrained(
    out_desc_array: *mut lldesc_t,
    buffer: *const c_void,
    size: c_int,
    max_desc_size: c_int,
    isrx: bool
)

Generate a linked list pointing to a (huge) buffer in an descriptor array.

The caller should ensure there is enough size to hold the array, by calling lldesc_get_required_num_constrained with the same max_desc_size argument.

@param[out] out_desc_array Output of a descriptor array, the head should be fed to the DMA. @param buffer Buffer for the descriptors to point to. @param size Size (or length for TX) of the buffer @param max_desc_size Maximum length of each descriptor @param isrx The RX DMA may require the buffer to be word-aligned, set to true for a RX link, otherwise false.