esp_idf_sys

Function mbedtls_asn1_get_bitstring_null

pub unsafe extern "C" fn mbedtls_asn1_get_bitstring_null(
    p: *mut *mut c_uchar,
    end: *const c_uchar,
    len: *mut usize,
) -> c_int

\brief Retrieve a bitstring ASN.1 tag without unused bits and its value. Updates the pointer to the beginning of the bit/octet string.

\param p On entry, \c *p points to the start of the ASN.1 element. On successful completion, \c *p points to the first byte of the content of the BIT STRING. On error, the value of \c *p is undefined. \param end End of data. \param len On success, \c *len is the length of the content in bytes.

\return 0 if successful. \return #MBEDTLS_ERR_ASN1_INVALID_DATA if the input starts with a valid BIT STRING with a nonzero number of unused bits. \return An ASN.1 error code if the input does not start with a valid ASN.1 BIT STRING.