esp_idf_sys

Function mbedtls_dhm_read_public

pub unsafe extern "C" fn mbedtls_dhm_read_public(
    ctx: *mut mbedtls_dhm_context,
    input: *const c_uchar,
    ilen: usize,
) -> c_int

\brief This function imports the raw public value of the peer.

\note In a TLS handshake, this is the how the server imports the Client’s public DHM key.

\param ctx The DHM context to use. This must be initialized and have its DHM parameters set, e.g. via mbedtls_dhm_set_group(). It may or may not already have generated its own private key. \param input The input buffer containing the \c G^Y value of the peer. This must be a readable buffer of size \p ilen Bytes. \param ilen The size of the input buffer \p input in Bytes.

\return \c 0 on success. \return An \c MBEDTLS_ERR_DHM_XXX error code on failure.