Function esp_idf_sys::mbedtls_ecc_group_to_psa

pub unsafe extern "C" fn mbedtls_ecc_group_to_psa(
    grpid: mbedtls_ecp_group_id,
    bits: *mut usize,
) -> psa_ecc_family_t

Convert an ECC curve identifier from the Mbed TLS encoding to PSA.

\note This function is provided solely for the convenience of Mbed TLS and may be removed at any time without notice.

\param grpid An Mbed TLS elliptic curve identifier (MBEDTLS_ECP_DP_xxx). \param[out] bits On success, the bit size of the curve.

\return The corresponding PSA elliptic curve identifier (PSA_ECC_FAMILY_xxx). \return \c 0 on failure (\p grpid is not recognized).