Function esp_idf_sys::mbedtls_ecp_curve_list

pub unsafe extern "C" fn mbedtls_ecp_curve_list(
) -> *const mbedtls_ecp_curve_info

\brief This function retrieves the information defined in mbedtls_ecp_curve_info() for all supported curves.

\note This function returns information about all curves supported by the library. Some curves may not be supported for all algorithms. Call mbedtls_ecdh_can_do() or mbedtls_ecdsa_can_do() to check if a curve is supported for ECDH or ECDSA.

\return A statically allocated array. The last entry is 0.