esp_idf_sys
Function mbedtls_entropy_add_source

pub unsafe extern "C" fn mbedtls_entropy_add_source(
    ctx: *mut mbedtls_entropy_context,
    f_source: mbedtls_entropy_f_source_ptr,
    p_source: *mut c_void,
    threshold: usize,
    strong: c_int,
) -> c_int

\brief Adds an entropy source to poll (Thread-safe if MBEDTLS_THREADING_C is enabled)

\param ctx Entropy context \param f_source Entropy function \param p_source Function data \param threshold Minimum required from source before entropy is released ( with mbedtls_entropy_func() ) (in bytes) \param strong MBEDTLS_ENTROPY_SOURCE_STRONG or MBEDTLS_ENTROPY_SOURCE_WEAK. At least one strong source needs to be added. Weaker sources (such as the cycle counter) can be used as a complement.

\return 0 if successful or MBEDTLS_ERR_ENTROPY_MAX_SOURCES