esp_idf_sys

Function mbedtls_hkdf

pub unsafe extern "C" fn mbedtls_hkdf(
    md: *const mbedtls_md_info_t,
    salt: *const c_uchar,
    salt_len: usize,
    ikm: *const c_uchar,
    ikm_len: usize,
    info: *const c_uchar,
    info_len: usize,
    okm: *mut c_uchar,
    okm_len: usize,
) -> c_int

\brief This is the HMAC-based Extract-and-Expand Key Derivation Function (HKDF).

\param md A hash function; md.size denotes the length of the hash function output in bytes. \param salt An optional salt value (a non-secret random value); if the salt is not provided, a string of all zeros of md.size length is used as the salt. \param salt_len The length in bytes of the optional \p salt. \param ikm The input keying material. \param ikm_len The length in bytes of \p ikm. \param info An optional context and application specific information string. This can be a zero-length string. \param info_len The length of \p info in bytes. \param okm The output keying material of \p okm_len bytes. \param okm_len The length of the output keying material in bytes. This must be less than or equal to 255 * md.size bytes.

\return 0 on success. \return #MBEDTLS_ERR_HKDF_BAD_INPUT_DATA when the parameters are invalid. \return An MBEDTLS_ERR_MD_* error for errors returned from the underlying MD layer.