Function esp_idf_sys::mbedtls_net_poll

pub unsafe extern "C" fn mbedtls_net_poll(
    ctx: *mut mbedtls_net_context, 
    rw: u32, 
    timeout: u32
) -> c_int

\brief Check and wait for the context to be ready for read/write

\note The current implementation of this function uses select() and returns an error if the file descriptor is \c FD_SETSIZE or greater.

\param ctx Socket to check \param rw Bitflag composed of MBEDTLS_NET_POLL_READ and MBEDTLS_NET_POLL_WRITE specifying the events to wait for: - If MBEDTLS_NET_POLL_READ is set, the function will return as soon as the net context is available for reading. - If MBEDTLS_NET_POLL_WRITE is set, the function will return as soon as the net context is available for writing. \param timeout Maximal amount of time to wait before returning, in milliseconds. If \c timeout is zero, the function returns immediately. If \c timeout is -1u, the function blocks potentially indefinitely.

\return Bitmask composed of MBEDTLS_NET_POLL_READ/WRITE on success or timeout, or a negative return code otherwise.