esp_idf_sys

Function mbedtls_pem_read_buffer

pub unsafe extern "C" fn mbedtls_pem_read_buffer(
    ctx: *mut mbedtls_pem_context,
    header: *const c_char,
    footer: *const c_char,
    data: *const c_uchar,
    pwd: *const c_uchar,
    pwdlen: usize,
    use_len: *mut usize,
) -> c_int

\brief Read a buffer for PEM information and store the resulting data into the specified context buffers.

\param ctx context to use \param header header string to seek and expect \param footer footer string to seek and expect \param data source data to look in (must be nul-terminated) \param pwd password for decryption (can be NULL) \param pwdlen length of password \param use_len destination for total length used from data buffer. It is set after header is correctly read, so unless you get MBEDTLS_ERR_PEM_BAD_INPUT_DATA or MBEDTLS_ERR_PEM_NO_HEADER_FOOTER_PRESENT, use_len is the length to skip.

\note Attempts to check password correctness by verifying if the decrypted text starts with an ASN.1 sequence of appropriate length

\note \c mbedtls_pem_free must be called on PEM context before the PEM context can be reused in another call to \c mbedtls_pem_read_buffer

\return 0 on success, or a specific PEM error code