esp_idf_sys

Function mbedtls_pk_check_pair

pub unsafe extern "C" fn mbedtls_pk_check_pair(
    pub_: *const mbedtls_pk_context,
    prv: *const mbedtls_pk_context,
    f_rng: mbedtls_f_rng_t,
    p_rng: *mut c_void,
) -> c_int

\brief Check if a public-private pair of keys matches.

\param pub Context holding a public key. \param prv Context holding a private (and public) key. \param f_rng RNG function, must not be \c NULL. \param p_rng RNG parameter

\return \c 0 on success (keys were checked and match each other). \return #MBEDTLS_ERR_PK_FEATURE_UNAVAILABLE if the keys could not be checked - in that case they may or may not match. \return #MBEDTLS_ERR_PK_BAD_INPUT_DATA if a context is invalid. \return Another non-zero value if the keys do not match.