esp_idf_sys

Function mbedtls_rsa_rsassa_pkcs1_v15_sign

pub unsafe extern "C" fn mbedtls_rsa_rsassa_pkcs1_v15_sign(
    ctx: *mut mbedtls_rsa_context,
    f_rng: mbedtls_f_rng_t,
    p_rng: *mut c_void,
    md_alg: mbedtls_md_type_t,
    hashlen: c_uint,
    hash: *const c_uchar,
    sig: *mut c_uchar,
) -> c_int

\brief This function performs a PKCS#1 v1.5 signature operation (RSASSA-PKCS1-v1_5-SIGN).

\param ctx The initialized RSA context to use. \param f_rng The RNG function. This is used for blinding and is mandatory; see mbedtls_rsa_private() for more. \param p_rng The RNG context to be passed to \p f_rng. This may be \c NULL if \p f_rng doesn’t need a context argument. \param md_alg The message-digest algorithm used to hash the original data. Use #MBEDTLS_MD_NONE for signing raw data. \param hashlen The length of the message digest or raw data in Bytes. If \p md_alg is not #MBEDTLS_MD_NONE, this must match the output length of the corresponding hash algorithm. \param hash The buffer holding the message digest or raw data. This must be a readable buffer of at least \p hashlen Bytes. \param sig The buffer to hold the signature. This must be a writable buffer of length \c ctx->len Bytes. For example, \c 256 Bytes for an 2048-bit RSA modulus. A buffer length of #MBEDTLS_MPI_MAX_SIZE is always safe.

\return \c 0 if the signing operation was successful. \return An \c MBEDTLS_ERR_RSA_XXX error code on failure.