esp_idf_sys
Function mbedtls_sha3_finish

pub unsafe extern "C" fn mbedtls_sha3_finish(
    ctx: *mut mbedtls_sha3_context,
    output: *mut u8,
    olen: usize,
) -> c_int

\brief This function finishes the SHA-3 operation, and writes the result to the output buffer.

\param ctx The SHA-3 context. This must be initialized and have a hash operation started. \param output The SHA-3 checksum result. This must be a writable buffer of length \c olen bytes. \param olen Defines the length of output buffer (in bytes). For SHA-3 224, SHA-3 256, SHA-3 384 and SHA-3 512 \c olen must equal to 28, 32, 48 and 64, respectively.

\return \c 0 on success. \return A negative error code on failure.