Function esp_idf_sys::mbedtls_sha512_ret

pub unsafe extern "C" fn mbedtls_sha512_ret(
    input: *const c_uchar,
    ilen: usize,
    output: *mut c_uchar,
    is384: c_int
) -> c_int

\brief This function calculates the SHA-512 or SHA-384 checksum of a buffer.

            The function allocates the context, performs the
            calculation, and frees the context.

            The SHA-512 result is calculated as
            output = SHA-512(input buffer).

\param input The buffer holding the input data. This must be a readable buffer of length \p ilen Bytes. \param ilen The length of the input data in Bytes. \param output The SHA-384 or SHA-512 checksum result. This must be a writable buffer of length \c 64 Bytes. \param is384 Determines which function to use. This must be either \c 0 for SHA-512, or \c 1 for SHA-384.

\note When \c MBEDTLS_SHA512_NO_SHA384 is defined, \p is384 must be \c 0, or the function will return #MBEDTLS_ERR_SHA512_BAD_INPUT_DATA.

\return \c 0 on success. \return A negative error code on failure.