Function esp_idf_sys::mbedtls_ssl_conf_ciphersuites_for_version

pub unsafe extern "C" fn mbedtls_ssl_conf_ciphersuites_for_version(
    conf: *mut mbedtls_ssl_config, 
    ciphersuites: *const c_int, 
    major: c_int, 
    minor: c_int
)

\brief Set the list of allowed ciphersuites and the preference order for a specific version of the protocol. (Only useful on the server side)

                 The ciphersuites array is not copied, and must remain
                 valid for the lifetime of the ssl_config.

\param conf SSL configuration \param ciphersuites 0-terminated list of allowed ciphersuites \param major Major version number (only MBEDTLS_SSL_MAJOR_VERSION_3 supported) \param minor Minor version number (MBEDTLS_SSL_MINOR_VERSION_0, MBEDTLS_SSL_MINOR_VERSION_1 and MBEDTLS_SSL_MINOR_VERSION_2, MBEDTLS_SSL_MINOR_VERSION_3 supported)

\note With DTLS, use MBEDTLS_SSL_MINOR_VERSION_2 for DTLS 1.0 and MBEDTLS_SSL_MINOR_VERSION_3 for DTLS 1.2