esp_idf_sys

Function mbedtls_ssl_conf_renegotiation_period

pub unsafe extern "C" fn mbedtls_ssl_conf_renegotiation_period(
    conf: *mut mbedtls_ssl_config,
    period: *const c_uchar,
)

\brief Set record counter threshold for periodic renegotiation. (Default: 2^48 - 1)

            Renegotiation is automatically triggered when a record
            counter (outgoing or incoming) crosses the defined
            threshold. The default value is meant to prevent the
            connection from being closed when the counter is about to
            reached its maximal value (it is not allowed to wrap).

            Lower values can be used to enforce policies such as "keys
            must be refreshed every N packets with cipher X".

            The renegotiation period can be disabled by setting
            conf->disable_renegotiation to
            MBEDTLS_SSL_RENEGOTIATION_DISABLED.

\note When the configured transport is MBEDTLS_SSL_TRANSPORT_DATAGRAM the maximum renegotiation period is 2^48 - 1, and for MBEDTLS_SSL_TRANSPORT_STREAM, the maximum renegotiation period is 2^64 - 1.

\param conf SSL configuration \param period The threshold value: a big-endian 64-bit number.