esp_idf_sys
Function mbedtls_ssl_get_max_in_record_payload

pub unsafe extern "C" fn mbedtls_ssl_get_max_in_record_payload(
    ssl: *const mbedtls_ssl_context,
) -> c_int

\brief Return the current maximum incoming record payload in bytes.

\note The logic to determine the maximum incoming record payload is version-specific. It takes into account various factors, such as the mbedtls_config.h setting \c MBEDTLS_SSL_IN_CONTENT_LEN, extensions such as the max fragment length extension or record size limit extension if used, and the current record expansion.

\sa mbedtls_ssl_set_mtu() \sa mbedtls_ssl_get_max_in_record_payload() \sa mbedtls_ssl_get_record_expansion()

\param ssl SSL context

\return Current maximum payload for an incoming record, or a negative error code.