esp_idf_sys

Function mbedtls_ssl_set_export_keys_cb

pub unsafe extern "C" fn mbedtls_ssl_set_export_keys_cb(
    ssl: *mut mbedtls_ssl_context,
    f_export_keys: mbedtls_ssl_export_keys_t,
    p_export_keys: *mut c_void,
)

\brief Configure a key export callback. (Default: none.)

     This API can be used for two purposes:
     - Debugging: Use this API to e.g. generate an NSSKeylog
       file and use it to inspect encrypted traffic in tools
       such as Wireshark.
     - Application-specific export: Use this API to implement
       key exporters, e.g. for EAP-TLS or DTLS-SRTP.

\param ssl The SSL context to which the export callback should be attached. \param f_export_keys The callback for the key export. \param p_export_keys The opaque context pointer to be passed to the callback \p f_export_keys.