esp_idf_sys

Function mbedtls_ssl_set_verify

pub unsafe extern "C" fn mbedtls_ssl_set_verify(
    ssl: *mut mbedtls_ssl_context,
    f_vrfy: Option<unsafe extern "C" fn(arg1: *mut c_void, arg2: *mut mbedtls_x509_crt, arg3: c_int, arg4: *mut u32) -> c_int>,
    p_vrfy: *mut c_void,
)

\brief Set a connection-specific verification callback (optional).

            If set, the provided verify callback is called for each
            certificate in the peer's CRT chain, including the trusted
            root. For more information, please see the documentation of
            \c mbedtls_x509_crt_verify().

\note This call is analogous to mbedtls_ssl_conf_verify() but binds the verification callback and context to an SSL context as opposed to an SSL configuration. If mbedtls_ssl_conf_verify() and mbedtls_ssl_set_verify() are both used, mbedtls_ssl_set_verify() takes precedence.

\param ssl The SSL context to use. \param f_vrfy The verification callback to use during CRT verification. \param p_vrfy The opaque context to be passed to the callback.