Function esp_idf_sys::multi_heap_register

pub unsafe extern "C" fn multi_heap_register(
    start: *mut c_void, 
    size: size_t
) -> multi_heap_handle_t

@brief Register a new heap for use

This function initialises a heap at the specified address, and returns a handle for future heap operations.

There is no equivalent function for deregistering a heap - if all blocks in the heap are free, you can immediately start using the memory for other purposes.

@param start Start address of the memory to use for a new heap. @param size Size (in bytes) of the new heap.

@return Handle of a new heap ready for use, or NULL if the heap region was too small to be initialised.