esp_idf_sys

Function multi_heap_set_lock

pub unsafe extern "C" fn multi_heap_set_lock(
    heap: multi_heap_handle_t,
    lock: *mut c_void,
)

@brief Associate a private lock pointer with a heap

The lock argument is supplied to the MULTI_HEAP_LOCK() and MULTI_HEAP_UNLOCK() macros, defined in multi_heap_platform.h.

The lock in question must be recursive.

When the heap is first registered, the associated lock is NULL.

@param heap Handle to a registered heap. @param lock Optional pointer to a locking structure to associate with this heap.