esp_idf_sys

Function os_mbuf_appendfrom

pub unsafe extern "C" fn os_mbuf_appendfrom(
    dst: *mut os_mbuf,
    src: *const os_mbuf,
    src_off: u16,
    len: u16,
) -> c_int

Reads data from one mbuf and appends it to another. On error, the specified data range may be partially appended. Neither mbuf is required to contain an mbuf packet header.

@param dst The mbuf to append to. @param src The mbuf to copy data from. @param src_off The absolute offset within the source mbuf chain to read from. @param len The number of bytes to append.

@return 0 on success; OS_EINVAL if the specified range extends beyond the end of the source mbuf chain.