esp_idf_sys

Function otLinkActiveScan

pub unsafe extern "C" fn otLinkActiveScan(
    aInstance: *mut otInstance,
    aScanChannels: u32,
    aScanDuration: u16,
    aCallback: otHandleActiveScanResult,
    aCallbackContext: *mut c_void,
) -> otError

Starts an IEEE 802.15.4 Active Scan

@param[in] aInstance A pointer to an OpenThread instance. @param[in] aScanChannels A bit vector indicating which channels to scan (e.g. OT_CHANNEL_11_MASK). @param[in] aScanDuration The time in milliseconds to spend scanning each channel. @param[in] aCallback A pointer to a function called on receiving a beacon or scan completes. @param[in] aCallbackContext A pointer to application-specific context.

@retval OT_ERROR_NONE Accepted the Active Scan request. @retval OT_ERROR_BUSY Already performing an Active Scan.