esp_idf_sys
Function otLinkFilterGetNextAddress

pub unsafe extern "C" fn otLinkFilterGetNextAddress(
    aInstance: *mut otInstance,
    aIterator: *mut otMacFilterIterator,
    aEntry: *mut otMacFilterEntry,
) -> otError

Gets an in-use address filter entry.

Is available when OPENTHREAD_CONFIG_MAC_FILTER_ENABLE configuration is enabled.

@param[in] aInstance A pointer to an OpenThread instance. @param[in,out] aIterator A pointer to the MAC filter iterator context. To get the first in-use address filter entry, it should be set to OT_MAC_FILTER_ITERATOR_INIT. MUST NOT be NULL. @param[out] aEntry A pointer to where the information is placed. MUST NOT be NULL.

@retval OT_ERROR_NONE Successfully retrieved an in-use address filter entry. @retval OT_ERROR_NOT_FOUND No subsequent entry exists.