esp_idf_sys

Function otPlatCryptoEcdsaGetPublicKey

pub unsafe extern "C" fn otPlatCryptoEcdsaGetPublicKey(
    aKeyPair: *const otPlatCryptoEcdsaKeyPair,
    aPublicKey: *mut otPlatCryptoEcdsaPublicKey,
) -> otError

Get the associated public key from the input context.

@param[in] aKeyPair A pointer to an ECDSA key-pair structure where the key-pair is stored. @param[out] aPublicKey A pointer to an ECDSA public key structure to store the public key.

@retval OT_ERROR_NONE Public key was retrieved successfully, and @p aBuffer is updated. @retval OT_ERROR_PARSE The key-pair DER format could not be parsed (invalid format). @retval OT_ERROR_INVALID_ARGS The @p aContext is NULL.