esp_idf_sys

Function otPlatCryptoEcdsaVerifyUsingKeyRef

pub unsafe extern "C" fn otPlatCryptoEcdsaVerifyUsingKeyRef(
    aKeyRef: otCryptoKeyRef,
    aHash: *const otPlatCryptoSha256Hash,
    aSignature: *const otPlatCryptoEcdsaSignature,
) -> otError

Use the keyref to verify the ECDSA signature of a hashed message.

@param[in] aKeyRef Key Reference to the slot where the key-pair is stored. @param[in] aHash A pointer to a SHA-256 hash structure where the hash value for signature verification is stored. @param[in] aSignature A pointer to an ECDSA signature structure where the signature value to be verified is stored.

@retval OT_ERROR_NONE The signature was verified successfully. @retval OT_ERROR_SECURITY The signature is invalid. @retval OT_ERROR_INVALID_ARGS The key or hash is invalid. @retval OT_ERROR_NO_BUFS Failed to allocate buffer for signature verification.

@note This API is only used by OT core when OPENTHREAD_CONFIG_PLATFORM_KEY_REFERENCES_ENABLE is enabled.