esp_idf_sys

Function otPlatRadioConfigureEnhAckProbing

pub unsafe extern "C" fn otPlatRadioConfigureEnhAckProbing(
    aInstance: *mut otInstance,
    aLinkMetrics: otLinkMetrics,
    aShortAddress: otShortAddress,
    aExtAddress: *const otExtAddress,
) -> otError

Enable/disable or update Enhanced-ACK Based Probing in radio for a specific Initiator.

After Enhanced-ACK Based Probing is configured by a specific Probing Initiator, the Enhanced-ACK sent to that node should include Vendor-Specific IE containing Link Metrics data. This function informs the radio to start/stop to collect Link Metrics data and include Vendor-Specific IE that containing the data in Enhanced-ACK sent to that Probing Initiator.

Regarding @p aExtAddress, this function assumes big-endian byte order. Note that this differs from otPlatRadioSetExtendedAddress(), otPlatRadioAddSrcMatchExtEntry(), and otPlatRadioClearSrcMatchExtEntry(), which use little-endian byte order for the Extended MAC address.

@param[in] aInstance The OpenThread instance structure. @param[in] aLinkMetrics This parameter specifies what metrics to query. Per spec 4.11.3.4.4.6, at most 2 metrics can be specified. The probing would be disabled if @p aLinkMetrics is bitwise 0. @param[in] aShortAddress The short address of the Probing Initiator. @param[in] aExtAddress The extended source address of the Probing Initiator. @p aExtAddress MUST NOT be NULL. The @p aExtAddress assumes big-endian byte order.

@retval OT_ERROR_NONE Successfully configured the Enhanced-ACK Based Probing. @retval OT_ERROR_INVALID_ARGS @p aExtAddress is NULL. @retval OT_ERROR_NOT_FOUND The Initiator indicated by @p aShortAddress is not found when trying to clear. @retval OT_ERROR_NO_BUFS No more Initiator can be supported. @retval OT_ERROR_NOT_IMPLEMENTED The feature is not implemented.