esp_idf_sys
Function otThreadGetMaxTimeInQueue

pub unsafe extern "C" fn otThreadGetMaxTimeInQueue(
    aInstance: *mut otInstance,
) -> u32

Gets the maximum time-in-queue for messages in the TX queue.

Requires OPENTHREAD_CONFIG_TX_QUEUE_STATISTICS_ENABLE.

The time-in-queue is tracked for direct transmissions only and is measured as the duration from when a message is added to the transmit queue until it is passed to the MAC layer for transmission or dropped.

The collected statistics can be reset by calling otThreadResetTimeInQueueStat().

@param[in] aInstance A pointer to an OpenThread instance.

@returns The maximum time-in-queue in milliseconds for all messages in the TX queue (so far).