Function esp_idf_sys::psa_verify_hash_abort

pub unsafe extern "C" fn psa_verify_hash_abort(
    operation: *mut psa_verify_hash_interruptible_operation_t,
) -> psa_status_t

\brief Abort a verify hash operation.

\warning This is a beta API, and thus subject to change at any point. It is not bound by the usual interface stability promises.

\note This function is the only function that clears the number of ops completed as part of the operation. Please ensure you copy this value via \c psa_verify_hash_get_num_ops() if required before calling.

\note Aborting an operation frees all associated resources except for the operation structure itself. Once aborted, the operation object can be reused for another operation by calling \c psa_verify_hash_start() again.

\note You may call this function any time after the operation object has been initialized. In particular, calling \c psa_verify_hash_abort() after the operation has already been terminated by a call to \c psa_verify_hash_abort() or psa_verify_hash_complete() is safe.

\param[in,out] operation Initialized verify hash operation.

\retval #PSA_SUCCESS The operation was aborted successfully.

\retval #PSA_ERROR_NOT_SUPPORTED \emptydescription \retval #PSA_ERROR_BAD_STATE The library has not been previously initialized by psa_crypto_init(). It is implementation-dependent whether a failure to initialize results in this error code.