Function esp_idf_sys::sdmmc_io_print_cis_info

pub unsafe extern "C" fn sdmmc_io_print_cis_info(
    buffer: *mut u8,
    buffer_size: usize,
    fp: *mut FILE,
) -> esp_err_t

Parse and print the CIS information of an SDIO card.

@note Not all the CIS codes and all kinds of tuples are supported. If you see some unresolved code, you can add the parsing of these code in sdmmc_io.c and contribute to the IDF through the Github repository.

         using sdmmc_card_init

@param buffer Buffer to parse @param buffer_size Size of the buffer. @param fp File pointer to print to, set to NULL to print to stdout.

@return - ESP_OK: on success - ESP_ERR_NOT_SUPPORTED: if the value from the card is not supported to be parsed. - ESP_ERR_INVALID_SIZE: if the CIS size fields are not correct.