esp_idf_sys
Function sdmmc_io_write_blocks

pub unsafe extern "C" fn sdmmc_io_write_blocks(
    card: *mut sdmmc_card_t,
    function: u32,
    addr: u32,
    src: *const c_void,
    size: usize,
) -> esp_err_t

Write blocks of data to an SDIO card using IO_RW_EXTENDED (CMD53)

This function performs write operation using CMD53 in block mode. For byte mode, see sdmmc_io_write_bytes.

By default OP Code is set (incrementing address). To send CMD53 without this bit, OR the argument addr with SDMMC_IO_FIXED_ADDR.

@param card pointer to card information structure previously initialized using sdmmc_card_init @param function IO function number @param addr byte address within IO function where writing starts @param src data to be written. Aligned to 4 byte boundary, and also cache line size if the buffer is behind the cache. @param size number of bytes to read, must be divisible by the card block size. @return - ESP_OK on success - ESP_ERR_INVALID_SIZE if size is not divisible by 512 bytes - One of the error codes from SDMMC host controller