esp_idf_sys

Function sdmmc_io_write_byte

pub unsafe extern "C" fn sdmmc_io_write_byte(
    card: *mut sdmmc_card_t,
    function: u32,
    reg: u32,
    in_byte: u8,
    out_byte: *mut u8,
) -> esp_err_t

Write one byte to an SDIO card using IO_RW_DIRECT (CMD52)

@param card pointer to card information structure previously initialized using sdmmc_card_init @param function IO function number @param reg byte address within IO function @param in_byte value to be written @param[out] out_byte if not NULL, receives new byte value read from the card (read-after-write). @return - ESP_OK on success - One of the error codes from SDMMC host controller