esp_idf_sys
Function sdspi_host_do_transaction

pub unsafe extern "C" fn sdspi_host_do_transaction(
    handle: sdspi_dev_handle_t,
    cmdinfo: *mut sdmmc_command_t,
) -> esp_err_t

@brief Send command to the card and get response

This function returns when command is sent and response is received, or data is transferred, or timeout occurs.

@note This function is not thread safe w.r.t. init/deinit functions, and bus width/clock speed configuration functions. Multiple tasks can call sdspi_host_do_transaction as long as other sdspi_host_* functions are not called.

@param handle Handle of the sdspi device @param cmdinfo pointer to structure describing command and data to transfer @return - ESP_OK on success - ESP_ERR_TIMEOUT if response or data transfer has timed out - ESP_ERR_INVALID_CRC if response or data transfer CRC check has failed - ESP_ERR_INVALID_RESPONSE if the card has sent an invalid response