esp_idf_sys
Function twai_receive

pub unsafe extern "C" fn twai_receive(
    message: *mut twai_message_t,
    ticks_to_wait: TickType_t,
) -> esp_err_t

@brief Receive a TWAI message

This function receives a message from the RX queue. The flags field of the message structure will indicate the type of message received. This function will block if there are no messages in the RX queue

@param[out] message Received message @param[in] ticks_to_wait Number of FreeRTOS ticks to block on RX queue

@warning The flags field of the received message should be checked to determine if the received message contains any data bytes.

@return - ESP_OK: Message successfully received from RX queue - ESP_ERR_TIMEOUT: Timed out waiting for message - ESP_ERR_INVALID_ARG: Arguments are invalid - ESP_ERR_INVALID_STATE: TWAI driver is not installed