Function esp_idf_sys::wifi_prov_mgr_is_provisioned

pub unsafe extern "C" fn wifi_prov_mgr_is_provisioned(
    provisioned: *mut bool,
) -> esp_err_t

@brief Checks if device is provisioned

This checks if Wi-Fi credentials are present on the NVS

The Wi-Fi credentials are assumed to be kept in the same NVS namespace as used by esp_wifi component

If one were to call esp_wifi_set_config() directly instead of going through the provisioning process, this function will still yield true (i.e. device will be found to be provisioned)

@note Calling wifi_prov_mgr_start_provisioning() automatically resets the provision state, irrespective of what the state was prior to making the call.

@param[out] provisioned True if provisioned, else false

@return

ESP_OK : Retrieved provision state successfully

ESP_FAIL : Wi-Fi not initialized

ESP_ERR_INVALID_ARG : Null argument supplied