Function esp_idf_sys::wifi_prov_mgr_set_app_info

pub unsafe extern "C" fn wifi_prov_mgr_set_app_info(
    label: *const c_char, 
    version: *const c_char, 
    capabilities: *mut *const c_char, 
    total_capabilities: size_t
) -> esp_err_t

@brief Set application version and capabilities in the JSON data returned by proto-ver endpoint

This function can be called multiple times, to specify information about the various application specific services running on the device, identified by unique labels.

The provisioning service itself registers an entry in the JSON data, by the label “prov”, containing only provisioning service version and capabilities. Application services should use a label other than “prov” so as not to overwrite this.

@note This must be called before executing wifi_prov_mgr_start_provisioning()

@param[in] label String indicating the application name.

@param[in] version String indicating the application version. There is no constraint on format.

@param[in] capabilities Array of strings with capabilities. These could be used by the client side app to know the application registered endpoint capabilities

@param[in] total_capabilities Size of capabilities array

@return

ESP_OK : Success

ESP_ERR_INVALID_STATE : Manager not initialized or provisioning service already started

ESP_ERR_NO_MEM : Failed to allocate memory for version string

ESP_ERR_INVALID_ARG : Null argument