esp_idf_sys

Function xRingbufferGetCurFreeSize

pub unsafe extern "C" fn xRingbufferGetCurFreeSize(
    xRingbuffer: RingbufHandle_t,
) -> usize

@brief Get current free size available for an item/data in the buffer

This gives the real time free space available for an item/data in the ring buffer. This represents the maximum size an item/data can have if it was currently sent to the ring buffer.

@note An empty no-split buffer has a max current free size for an item that is limited to ((buffer_size/2)-header_size). See API reference for xRingbufferGetMaxItemSize().

@param[in] xRingbuffer Ring buffer to query

@return Current free size, in bytes, available for an entry