Function esp_idf_sys::xTaskAbortDelay

pub unsafe extern "C" fn xTaskAbortDelay(
    xTask: TaskHandle_t
) -> BaseType_t

INCLUDE_xTaskAbortDelay must be defined as 1 in FreeRTOSConfig.h for this function to be available.

A task will enter the Blocked state when it is waiting for an event. The event it is waiting for can be a temporal event (waiting for a time), such as when vTaskDelay() is called, or an event on an object, such as when xQueueReceive() or ulTaskNotifyTake() is called. If the handle of a task that is in the Blocked state is used in a call to xTaskAbortDelay() then the task will leave the Blocked state, and return from whichever function call placed the task into the Blocked state.

@param xTask The handle of the task to remove from the Blocked state.

@return If the task referenced by xTask was not in the Blocked state then pdFAIL is returned. Otherwise pdPASS is returned.

\defgroup xTaskAbortDelay xTaskAbortDelay \ingroup TaskCtrl