Struct esp_idf_sys::EspError

pub struct EspError(/* private fields */);

A wrapped esp_err_t to check if an error occurred.

An esp_err_t is returned from most esp-idf APIs as a status code. If it is equal to ESP_OK it means no error occurred.

Implementations

impl EspError

pub const fn from(error: esp_err_t) -> Option<Self>

Wrap an esp_err_t, return [Some] if error is not ESP_OK.

pub const fn from_non_zero(error: NonZeroI32) -> Self

Wrap a [NonZeroI32]. Since ESP_OK is 0, this can never fail;

pub const fn from_infallible<const E: esp_err_t>() -> Self

Wrap an esp_err_t, throw a compile time error if error is ESP_OK.

pub fn check_and_return<T>(error: esp_err_t, value: T) -> Result<T, Self>

Convert error into a [Result] with Ok(value) if no error occurred.

If error is ESP_OK return [Ok] of value otherwise return [Err] of wrapped error.

pub fn convert(error: esp_err_t) -> Result<(), Self>

Convert error into a [Result] with Ok(()) if not error occurred..

If error equals to ESP_OK return [Ok], otherwise return [Err] with the wrapped esp_err_t.

pub fn panic(&self)

Panic with a specific error message of the contained esp_err_t.

pub fn code(&self) -> esp_err_t

Get the wrapped esp_err_t.