esp_idf_sys

Struct ble_hs_cfg

#[repr(C)]
pub struct ble_hs_cfg {

    pub gatts_register_cb: ble_gatt_register_fn,
    pub gatts_register_arg: *mut c_void,
    pub sm_io_cap: u8,
    pub _bitfield_align_1: [u8; 0],
    pub _bitfield_1: __BindgenBitfieldUnit<[u8; 1]>,
    pub sm_sec_lvl: u8,
    pub _bitfield_align_2: [u8; 0],
    pub _bitfield_2: __BindgenBitfieldUnit<[u8; 1]>,
    pub eatt: u8,
    pub sm_our_key_dist: u8,
    pub sm_their_key_dist: u8,
    pub gatt_use_cache: u8,
    pub reset_cb: ble_hs_reset_fn,
    pub sync_cb: ble_hs_sync_fn,
    pub store_gen_key_cb: ble_store_gen_key_fn,
    pub store_read_cb: ble_store_read_fn,
    pub store_write_cb: ble_store_write_fn,
    pub store_delete_cb: ble_store_delete_fn,
    pub store_status_cb: ble_store_status_fn,
    pub store_status_arg: *mut c_void,

}

@brief Bluetooth Host main configuration structure

Those can be used by application to configure stack.

The only reason Security Manager (sm_ members) is configurable at runtime is to simplify security testing. Defaults for those are configured by selecting proper options in application’s syscfg.

Fields
gatts_register_cb: ble_gatt_register_fn

An optional callback that gets executed upon registration of each GATT resource (service, characteristic, or descriptor).

gatts_register_arg: *mut c_void

An optional argument that gets passed to the GATT registration callback.

sm_io_cap: u8

Security Manager Local Input Output Capabilities

_bitfield_align_1: [u8; 0]_bitfield_1: __BindgenBitfieldUnit<[u8; 1]>sm_sec_lvl: u8

@brief Security Manager - Security Mode 1 Level for GATT related operations

Possible values are: 0: Default value, ignored 1: No security 2: Unauthenticated pairing with encryption 3. Authenticated pairing with encryption 4. Authenticated LE Secure Connections pairing with encryption using a 128-bit strength encryption key

_bitfield_align_2: [u8; 0]_bitfield_2: __BindgenBitfieldUnit<[u8; 1]>eatt: u8

@brief Enable/Disable Enhanced ATT Support

Primarily used to enable EATT behaviour; denotes the number of eatt channels. Set to 0 to disable eatt.

Default value is CONFIG_BT_NIMBLE_EATT_CHAN_NUM.

sm_our_key_dist: u8

@brief Security Manager Local Key Distribution Mask

sm_their_key_dist: u8

@brief Security Manager Remote Key Distribution Mask

gatt_use_cache: u8

@brief Weather to use GATT caching or not for discovery operations

reset_cb: ble_hs_reset_fn

@brief Stack reset callback

This callback is executed when the host resets itself and the controller due to fatal error.

sync_cb: ble_hs_sync_fn

@brief Stack sync callback

This callback is executed when the host and controller become synced. This happens at startup and after a reset.

store_gen_key_cb: ble_store_gen_key_fn

Callback to handle generation of security keys

store_read_cb: ble_store_read_fn

Storage Read callback handles read of security material

store_write_cb: ble_store_write_fn

Storage Write callback handles write of security material

store_delete_cb: ble_store_delete_fn

Storage Delete callback handles deletion of security material

store_status_cb: ble_store_status_fn

@brief Storage Status callback.

This callback gets executed when a persistence operation cannot be performed or a persistence failure is imminent. For example, if is insufficient storage capacity for a record to be persisted, this function gets called to give the application the opportunity to make room.

store_status_arg: *mut c_void

An optional argument that gets passed to the storage status callback.

Implementations

impl ble_hs_cfg

pub fn sm_oob_data_flag(&self) -> c_uint

pub fn set_sm_oob_data_flag(&mut self, val: c_uint)

pub unsafe fn sm_oob_data_flag_raw(this: *const Self) -> c_uint

pub unsafe fn set_sm_oob_data_flag_raw(this: *mut Self, val: c_uint)

pub fn sm_bonding(&self) -> c_uint

pub fn set_sm_bonding(&mut self, val: c_uint)

pub unsafe fn sm_bonding_raw(this: *const Self) -> c_uint

pub unsafe fn set_sm_bonding_raw(this: *mut Self, val: c_uint)

pub fn sm_mitm(&self) -> c_uint

pub fn set_sm_mitm(&mut self, val: c_uint)

pub unsafe fn sm_mitm_raw(this: *const Self) -> c_uint

pub unsafe fn set_sm_mitm_raw(this: *mut Self, val: c_uint)

pub fn sm_sc(&self) -> c_uint

pub fn set_sm_sc(&mut self, val: c_uint)

pub unsafe fn sm_sc_raw(this: *const Self) -> c_uint

pub unsafe fn set_sm_sc_raw(this: *mut Self, val: c_uint)

pub fn sm_sc_only(&self) -> c_uint

pub fn set_sm_sc_only(&mut self, val: c_uint)

pub unsafe fn sm_sc_only_raw(this: *const Self) -> c_uint

pub unsafe fn set_sm_sc_only_raw(this: *mut Self, val: c_uint)

pub fn new_bitfield_1( sm_oob_data_flag: c_uint, sm_bonding: c_uint, sm_mitm: c_uint, sm_sc: c_uint, sm_sc_only: c_uint, ) -> __BindgenBitfieldUnit<[u8; 1]>

pub fn sm_keypress(&self) -> c_uint

pub fn set_sm_keypress(&mut self, val: c_uint)

pub unsafe fn sm_keypress_raw(this: *const Self) -> c_uint

pub unsafe fn set_sm_keypress_raw(this: *mut Self, val: c_uint)

pub fn new_bitfield_2(sm_keypress: c_uint) -> __BindgenBitfieldUnit<[u8; 1]>