Struct esp_idf_sys::esp_image_header_t

#[repr(C, packed)]
pub struct esp_image_header_t {
    pub magic: u8,
    pub segment_count: u8,
    pub spi_mode: u8,
    pub _bitfield_align_1: [u8; 0],
    pub _bitfield_1: __BindgenBitfieldUnit<[u8; 1]>,
    pub entry_addr: u32,
    pub wp_pin: u8,
    pub spi_pin_drv: [u8; 3],
    pub chip_id: esp_chip_id_t,
    pub min_chip_rev: u8,
    pub reserved: [u8; 8],
    pub hash_appended: u8,
}

@brief Main header of binary image

Fields
magic: u8

< Magic word ESP_IMAGE_HEADER_MAGIC

segment_count: u8

< Count of memory segments

spi_mode: u8

< flash read mode (esp_image_spi_mode_t as uint8_t)

_bitfield_align_1: [u8; 0]_bitfield_1: __BindgenBitfieldUnit<[u8; 1]>entry_addr: u32

< Entry address

wp_pin: u8

< WP pin when SPI pins set via efuse (read by ROM bootloader, the IDF bootloader uses software to configure the WP pin and sets this field to 0xEE=disabled)

spi_pin_drv: [u8; 3]

< Drive settings for the SPI flash pins (read by ROM bootloader)

chip_id: esp_chip_id_t

< Chip identification number

min_chip_rev: u8

< Minimum chip revision supported by image

reserved: [u8; 8]

< Reserved bytes in additional header space, currently unused

hash_appended: u8

< If 1, a SHA256 digest “simple hash” (of the entire image) is appended after the checksum. Included in image length. This digest is separate to secure boot and only used for detecting corruption. For secure boot signed images, the signature is appended after this (and the simple hash is included in the signed data).

Implementations

impl esp_image_header_t

pub fn spi_speed(&self) -> u8

pub fn set_spi_speed(&mut self, val: u8)

pub fn spi_size(&self) -> u8

pub fn set_spi_size(&mut self, val: u8)

pub fn new_bitfield_1(
spi_speed: u8,
spi_size: u8
) -> __BindgenBitfieldUnit<[u8; 1]>