Struct esp_idf_sys::esp_mqtt_client_config_t

#[repr(C)]
pub struct esp_mqtt_client_config_t {

    pub event_handle: mqtt_event_callback_t,
    pub event_loop_handle: esp_event_loop_handle_t,
    pub host: *const c_char,
    pub uri: *const c_char,
    pub port: u32,
    pub client_id: *const c_char,
    pub username: *const c_char,
    pub password: *const c_char,
    pub lwt_topic: *const c_char,
    pub lwt_msg: *const c_char,
    pub lwt_qos: c_int,
    pub lwt_retain: c_int,
    pub lwt_msg_len: c_int,
    pub disable_clean_session: c_int,
    pub keepalive: c_int,
    pub disable_auto_reconnect: bool,
    pub user_context: *mut c_void,
    pub task_prio: c_int,
    pub task_stack: c_int,
    pub buffer_size: c_int,
    pub cert_pem: *const c_char,
    pub cert_len: size_t,
    pub client_cert_pem: *const c_char,
    pub client_cert_len: size_t,
    pub client_key_pem: *const c_char,
    pub client_key_len: size_t,
    pub transport: esp_mqtt_transport_t,
    pub refresh_connection_after_ms: c_int,
    pub psk_hint_key: *const psk_key_hint,
    pub use_global_ca_store: bool,
    pub crt_bundle_attach: Option<unsafe extern "C" fn(conf: *mut c_void) -> esp_err_t>,
    pub reconnect_timeout_ms: c_int,
    pub alpn_protos: *mut *const c_char,
    pub clientkey_password: *const c_char,
    pub clientkey_password_len: c_int,
    pub protocol_ver: esp_mqtt_protocol_ver_t,
    pub out_buffer_size: c_int,
    pub skip_cert_common_name_check: bool,
    pub use_secure_element: bool,
    pub ds_data: *mut c_void,
    pub network_timeout_ms: c_int,
    pub disable_keepalive: bool,
    pub path: *const c_char,

}

MQTT client configuration structure

Fields
event_handle: mqtt_event_callback_t

< handle for MQTT events as a callback in legacy mode

event_loop_handle: esp_event_loop_handle_t

< handle for MQTT event loop library

host: *const c_char

< MQTT server domain (ipv4 as string)

uri: *const c_char

< Complete MQTT broker URI

port: u32

< MQTT server port

client_id: *const c_char

< default client id is ESP32_%CHIPID% where %CHIPID% are last 3 bytes of MAC address in hex format

username: *const c_char

< MQTT username

password: *const c_char

< MQTT password

lwt_topic: *const c_char

< LWT (Last Will and Testament) message topic (NULL by default)

lwt_msg: *const c_char

< LWT message (NULL by default)

lwt_qos: c_int

< LWT message qos

lwt_retain: c_int

< LWT retained message flag

lwt_msg_len: c_int

< LWT message length

disable_clean_session: c_int

< mqtt clean session, default clean_session is true

keepalive: c_int

< mqtt keepalive, default is 120 seconds

disable_auto_reconnect: bool

< this mqtt client will reconnect to server (when errors/disconnect). Set disable_auto_reconnect=true to disable

user_context: *mut c_void

< pass user context to this option, then can receive that context in event->user_context

task_prio: c_int

< MQTT task priority, default is 5, can be changed in make menuconfig

task_stack: c_int

< MQTT task stack size, default is 6144 bytes, can be changed in make menuconfig

buffer_size: c_int

< size of MQTT send/receive buffer, default is 1024 (only receive buffer size if out_buffer_size defined)

cert_pem: *const c_char

< Pointer to certificate data in PEM or DER format for server verify (with SSL), default is NULL, not required to verify the server. PEM-format must have a terminating NULL-character. DER-format requires the length to be passed in cert_len.

cert_len: size_t

< Length of the buffer pointed to by cert_pem. May be 0 for null-terminated pem

client_cert_pem: *const c_char

< Pointer to certificate data in PEM or DER format for SSL mutual authentication, default is NULL, not required if mutual authentication is not needed. If it is not NULL, also client_key_pem has to be provided. PEM-format must have a terminating NULL-character. DER-format requires the length to be passed in client_cert_len.

client_cert_len: size_t

< Length of the buffer pointed to by client_cert_pem. May be 0 for null-terminated pem

client_key_pem: *const c_char

< Pointer to private key data in PEM or DER format for SSL mutual authentication, default is NULL, not required if mutual authentication is not needed. If it is not NULL, also client_cert_pem has to be provided. PEM-format must have a terminating NULL-character. DER-format requires the length to be passed in client_key_len

client_key_len: size_t

< Length of the buffer pointed to by client_key_pem. May be 0 for null-terminated pem

transport: esp_mqtt_transport_t

< overrides URI transport

refresh_connection_after_ms: c_int

< Refresh connection after this value (in milliseconds)

psk_hint_key: *const psk_key_hint

< Pointer to PSK struct defined in esp_tls.h to enable PSK authentication (as alternative to certificate verification). If not NULL and server/client certificates are NULL, PSK is enabled

use_global_ca_store: bool

< Use a global ca_store for all the connections in which this bool is set.

crt_bundle_attach: Option<unsafe extern "C" fn(conf: *mut c_void) -> esp_err_t>

< Pointer to ESP x509 Certificate Bundle attach function for the usage of certification bundles in mqtts

reconnect_timeout_ms: c_int

< Reconnect to the broker after this value in miliseconds if auto reconnect is not disabled (defaults to 10s)

alpn_protos: *mut *const c_char

< NULL-terminated list of supported application protocols to be used for ALPN

clientkey_password: *const c_char

< Client key decryption password string

clientkey_password_len: c_int

< String length of the password pointed to by clientkey_password

protocol_ver: esp_mqtt_protocol_ver_t

< MQTT protocol version used for connection, defaults to value from menuconfig

out_buffer_size: c_int

< size of MQTT output buffer. If not defined, both output and input buffers have the same size defined as buffer_size

skip_cert_common_name_check: bool

< Skip any validation of server certificate CN field, this reduces the security of TLS and makes the mqtt client susceptible to MITM attacks

use_secure_element: bool

< enable secure element for enabling SSL connection

ds_data: *mut c_void

< carrier of handle for digital signature parameters

network_timeout_ms: c_int

< Abort network operation if it is not completed after this value, in milliseconds (defaults to 10s)

disable_keepalive: bool

< Set disable_keepalive=true to turn off keep-alive mechanism, false by default (keepalive is active by default). Note: setting the config value keepalive to 0 doesn’t disable keepalive feature, but uses a default keepalive period

path: *const c_char

< Path in the URI