esp_idf_sys
Struct i2s_driver_config_t

#[repr(C)]
pub struct i2s_driver_config_t {

    pub mode: i2s_mode_t,
    pub sample_rate: u32,
    pub bits_per_sample: i2s_bits_per_sample_t,
    pub channel_format: i2s_channel_fmt_t,
    pub communication_format: i2s_comm_format_t,
    pub intr_alloc_flags: c_int,
    pub __bindgen_anon_1: i2s_driver_config_t__bindgen_ty_1,
    pub __bindgen_anon_2: i2s_driver_config_t__bindgen_ty_2,
    pub use_apll: bool,
    pub tx_desc_auto_clear: bool,
    pub fixed_mclk: c_int,
    pub mclk_multiple: i2s_mclk_multiple_t,
    pub bits_per_chan: i2s_bits_per_chan_t,
    pub chan_mask: i2s_channel_t,
    pub total_chan: u32,
    pub left_align: bool,
    pub big_edin: bool,
    pub bit_order_msb: bool,
    pub skip_msk: bool,

}

@brief I2S driver configuration parameters

Fields
mode: i2s_mode_t

< I2S work mode

sample_rate: u32

< I2S sample rate

bits_per_sample: i2s_bits_per_sample_t

< I2S sample bits in one channel

channel_format: i2s_channel_fmt_t

< I2S channel format.

communication_format: i2s_comm_format_t

< I2S communication format

intr_alloc_flags: c_int

< Flags used to allocate the interrupt. One or multiple (ORred) ESP_INTR_FLAG_* values. See esp_intr_alloc.h for more info

__bindgen_anon_1: i2s_driver_config_t__bindgen_ty_1__bindgen_anon_2: i2s_driver_config_t__bindgen_ty_2use_apll: bool

< I2S using APLL as main I2S clock, enable it to get accurate clock

tx_desc_auto_clear: bool

< I2S auto clear tx descriptor if there is underflow condition (helps in avoiding noise in case of data unavailability)

fixed_mclk: c_int

< I2S using fixed MCLK output. If use_apll = true and fixed_mclk > 0, then the clock output for i2s is fixed and equal to the fixed_mclk value. If fixed_mclk set, mclk_multiple wonโt take effect

mclk_multiple: i2s_mclk_multiple_t

< The multiple of I2S master clock(MCLK) to sample rate

bits_per_chan: i2s_bits_per_chan_t

< I2S total bits in one channel๏ผ only take effect when larger than โbits_per_sampleโ, default โ0โ means equal to โbits_per_sampleโ

chan_mask: i2s_channel_t

< I2S active channel bit mask, set value in i2s_channel_t to enable specific channel, the bit map of active channel can not exceed (0x1<<total_chan).

total_chan: u32

< I2S Total number of channels. If it is smaller than the biggest active channel number, it will be set to this number automatically.

left_align: bool

< Set to enable left alignment

big_edin: bool

< Set to enable big endian

bit_order_msb: bool

< Set to enable msb order

skip_msk: bool

< Set to enable skip mask. If it is enabled, only the data of the enabled channels will be sent, otherwise all data stored in DMA TX buffer will be sent