Struct esp_idf_sys::i2s_pdm_tx_slot_config_t

#[repr(C)]
pub struct i2s_pdm_tx_slot_config_t {

    pub data_bit_width: i2s_data_bit_width_t,
    pub slot_bit_width: i2s_slot_bit_width_t,
    pub slot_mode: i2s_slot_mode_t,
    pub sd_prescale: u32,
    pub sd_scale: i2s_pdm_sig_scale_t,
    pub hp_scale: i2s_pdm_sig_scale_t,
    pub lp_scale: i2s_pdm_sig_scale_t,
    pub sinc_scale: i2s_pdm_sig_scale_t,
    pub line_mode: i2s_pdm_tx_line_mode_t,
    pub hp_en: bool,
    pub hp_cut_off_freq_hz: f32,
    pub sd_dither: u32,
    pub sd_dither2: u32,

}

@brief I2S slot configuration for pdm tx mode

Fields
data_bit_width: i2s_data_bit_width_t

< I2S sample data bit width (valid data bits per sample), only support 16 bits for PDM mode

slot_bit_width: i2s_slot_bit_width_t

< I2S slot bit width (total bits per slot), only support 16 bits for PDM mode

slot_mode: i2s_slot_mode_t

< Set mono or stereo mode with I2S_SLOT_MODE_MONO or I2S_SLOT_MODE_STEREO For PDM TX mode, mono means the data buffer only contains one slot data, Stereo means the data buffer contains two slots data

sd_prescale: u32

< Sigma-delta filter prescale

sd_scale: i2s_pdm_sig_scale_t

< Sigma-delta filter scaling value

hp_scale: i2s_pdm_sig_scale_t

< High pass filter scaling value

lp_scale: i2s_pdm_sig_scale_t

< Low pass filter scaling value

sinc_scale: i2s_pdm_sig_scale_t

< Sinc filter scaling value

line_mode: i2s_pdm_tx_line_mode_t

< PDM TX line mode, one-line codec, one-line dac, two-line dac mode can be selected

hp_en: bool

< High pass filter enable

hp_cut_off_freq_hz: f32

< High pass filter cut-off frequency, range 23.3Hz ~ 185Hz, see cut-off frequency sheet above

sd_dither: u32

< Sigma-delta filter dither

sd_dither2: u32

< Sigma-delta filter dither2