esp_idf_sys

Struct mbedtls_ccm_context

#[repr(C)]
pub struct mbedtls_ccm_context {
    pub private_y: [c_uchar; 16],
    pub private_ctr: [c_uchar; 16],
    pub private_plaintext_len: usize,
    pub private_add_len: usize,
    pub private_tag_len: usize,
    pub private_processed: usize,
    pub private_q: c_uint,
    pub private_mode: c_uint,
    pub private_cipher_ctx: mbedtls_cipher_context_t,
    pub private_state: c_int,
}

\brief The CCM context-type definition. The CCM context is passed to the APIs called.

Fields
private_y: [c_uchar; 16]

< The Y working buffer

private_ctr: [c_uchar; 16]

< The counter buffer

private_plaintext_len: usize

< Total plaintext length

private_add_len: usize

< Total authentication data length

private_tag_len: usize

< Total tag length

private_processed: usize

< Track how many bytes of input data were processed (chunked input). Used independently for both auth data and plaintext/ciphertext. This variable is set to zero after auth data input is finished.

private_q: c_uint

< The Q working value

private_mode: c_uint

< The operation to perform: #MBEDTLS_CCM_ENCRYPT or #MBEDTLS_CCM_DECRYPT or #MBEDTLS_CCM_STAR_ENCRYPT or #MBEDTLS_CCM_STAR_DECRYPT.

private_cipher_ctx: mbedtls_cipher_context_t

< The cipher context used.

private_state: c_int

< Working value holding context’s state. Used for chunked data input