Struct esp_idf_sys::mbedtls_cipher_info_t

#[repr(C)]
pub struct mbedtls_cipher_info_t {
    pub private_name: *const c_char,
    pub _bitfield_align_1: [u8; 0],
    pub _bitfield_1: __BindgenBitfieldUnit<[u8; 4]>,
}

Cipher information. Allows calling cipher functions in a generic way.

\note The library does not support custom cipher info structures, only built-in structures returned by the functions mbedtls_cipher_info_from_string(), mbedtls_cipher_info_from_type(), mbedtls_cipher_info_from_values(), mbedtls_cipher_info_from_psa().

\note Some fields store a value that has been right-shifted to save code-size, so should not be used directly. The accessor functions adjust for this and return the “natural” value.

Fields
private_name: *const c_char_bitfield_align_1: [u8; 0]_bitfield_1: __BindgenBitfieldUnit<[u8; 4]>
Implementations

impl mbedtls_cipher_info_t

pub fn private_block_size(&self) -> c_uint

pub fn set_private_block_size(&mut self, val: c_uint)

pub fn private_iv_size(&self) -> c_uint

pub fn set_private_iv_size(&mut self, val: c_uint)

pub fn private_key_bitlen(&self) -> c_uint

pub fn set_private_key_bitlen(&mut self, val: c_uint)

pub fn private_mode(&self) -> c_uint

pub fn set_private_mode(&mut self, val: c_uint)

pub fn private_type(&self) -> c_uint

pub fn set_private_type(&mut self, val: c_uint)

pub fn private_flags(&self) -> c_uint

pub fn set_private_flags(&mut self, val: c_uint)

pub fn private_base_idx(&self) -> c_uint

pub fn set_private_base_idx(&mut self, val: c_uint)

pub fn new_bitfield_1( private_block_size: c_uint, private_iv_size: c_uint, private_key_bitlen: c_uint, private_mode: c_uint, private_type: c_uint, private_flags: c_uint, private_base_idx: c_uint, ) -> __BindgenBitfieldUnit<[u8; 4]>