esp_idf_sys

Struct mbedtls_mpi

#[repr(C)]
pub struct mbedtls_mpi {
    pub private_p: *mut mbedtls_mpi_uint,
    pub private_s: c_short,
    pub private_n: c_ushort,
}

\brief MPI structure

Fields
private_p: *mut mbedtls_mpi_uint

Pointer to limbs.

This may be \c NULL if \c n is 0.

private_s: c_short

Sign: -1 if the mpi is negative, 1 otherwise.

The number 0 must be represented with s = +1. Although many library functions treat all-limbs-zero as equivalent to a valid representation of 0 regardless of the sign bit, there are exceptions, so bignum functions and external callers must always set \c s to +1 for the number zero.

Note that this implies that calloc() or ... = {0} does not create a valid MPI representation. You must call mbedtls_mpi_init().

private_n: c_ushort

Total number of limbs in \c p.