Struct esp_idf_sys::mesh_cfg_t

#[repr(C)]
pub struct mesh_cfg_t {
    pub channel: u8,
    pub allow_channel_switch: bool,
    pub mesh_id: mesh_addr_t,
    pub router: mesh_router_t,
    pub mesh_ap: mesh_ap_cfg_t,
    pub crypto_funcs: *const mesh_crypto_funcs_t,
}

@brief Mesh initialization configuration

Fields
channel: u8

< channel, the mesh network on

allow_channel_switch: bool

< if this value is set, when “fail” (mesh_attempts_t) times is reached, device will change to a full channel scan for a network that could join. The default value is false.

mesh_id: mesh_addr_t

< mesh network identification

router: mesh_router_t

< router configuration

mesh_ap: mesh_ap_cfg_t

< mesh softAP configuration

crypto_funcs: *const mesh_crypto_funcs_t

< crypto functions