esp_idf_sys
Struct otBufferInfo

#[repr(C)]
pub struct otBufferInfo {
    pub mTotalBuffers: u16,
    pub mFreeBuffers: u16,
    pub mMaxUsedBuffers: u16,
    pub m6loSendQueue: otMessageQueueInfo,
    pub m6loReassemblyQueue: otMessageQueueInfo,
    pub mIp6Queue: otMessageQueueInfo,
    pub mMplQueue: otMessageQueueInfo,
    pub mMleQueue: otMessageQueueInfo,
    pub mCoapQueue: otMessageQueueInfo,
    pub mCoapSecureQueue: otMessageQueueInfo,
    pub mApplicationCoapQueue: otMessageQueueInfo,
}

Represents the message buffer information for different queues used by OpenThread stack.

Fields
mTotalBuffers: u16

< The total number of buffers in the messages pool (0xffff if unknown).

mFreeBuffers: u16

< The number of free buffers (0xffff if unknown).

mMaxUsedBuffers: u16

The maximum number of used buffers at the same time since OT stack initialization or last call to otMessageResetBufferInfo().

m6loSendQueue: otMessageQueueInfo

< Info about 6LoWPAN send queue.

m6loReassemblyQueue: otMessageQueueInfo

< Info about 6LoWPAN reassembly queue.

mIp6Queue: otMessageQueueInfo

< Info about IPv6 send queue.

mMplQueue: otMessageQueueInfo

< Info about MPL send queue.

mMleQueue: otMessageQueueInfo

< Info about MLE delayed message queue.

mCoapQueue: otMessageQueueInfo

< Info about CoAP/TMF send queue.

mCoapSecureQueue: otMessageQueueInfo

< Info about CoAP secure send queue.

mApplicationCoapQueue: otMessageQueueInfo

< Info about application CoAP send queue.