esp_idf_sys
Struct otRadioCoexMetrics

#[repr(C)]
pub struct otRadioCoexMetrics {

    pub mNumGrantGlitch: u32,
    pub mNumTxRequest: u32,
    pub mNumTxGrantImmediate: u32,
    pub mNumTxGrantWait: u32,
    pub mNumTxGrantWaitActivated: u32,
    pub mNumTxGrantWaitTimeout: u32,
    pub mNumTxGrantDeactivatedDuringRequest: u32,
    pub mNumTxDelayedGrant: u32,
    pub mAvgTxRequestToGrantTime: u32,
    pub mNumRxRequest: u32,
    pub mNumRxGrantImmediate: u32,
    pub mNumRxGrantWait: u32,
    pub mNumRxGrantWaitActivated: u32,
    pub mNumRxGrantWaitTimeout: u32,
    pub mNumRxGrantDeactivatedDuringRequest: u32,
    pub mNumRxDelayedGrant: u32,
    pub mAvgRxRequestToGrantTime: u32,
    pub mNumRxGrantNone: u32,
    pub mStopped: bool,

}

Represents radio coexistence metrics.

Fields
mNumGrantGlitch: u32

< Number of grant glitches.

mNumTxRequest: u32

< Number of tx requests.

mNumTxGrantImmediate: u32

< Number of tx requests while grant was active.

mNumTxGrantWait: u32

< Number of tx requests while grant was inactive.

mNumTxGrantWaitActivated: u32

< Number of tx requests while grant was inactive that were ultimately granted.

mNumTxGrantWaitTimeout: u32

< Number of tx requests while grant was inactive that timed out.

mNumTxGrantDeactivatedDuringRequest: u32

< Number of tx that were in progress when grant was deactivated.

mNumTxDelayedGrant: u32

< Number of tx requests that were not granted within 50us.

mAvgTxRequestToGrantTime: u32

< Average time in usec from tx request to grant.

mNumRxRequest: u32

< Number of rx requests.

mNumRxGrantImmediate: u32

< Number of rx requests while grant was active.

mNumRxGrantWait: u32

< Number of rx requests while grant was inactive.

mNumRxGrantWaitActivated: u32

< Number of rx requests while grant was inactive that were ultimately granted.

mNumRxGrantWaitTimeout: u32

< Number of rx requests while grant was inactive that timed out.

mNumRxGrantDeactivatedDuringRequest: u32

< Number of rx that were in progress when grant was deactivated.

mNumRxDelayedGrant: u32

< Number of rx requests that were not granted within 50us.

mAvgRxRequestToGrantTime: u32

< Average time in usec from rx request to grant.

mNumRxGrantNone: u32

< Number of rx requests that completed without receiving grant.

mStopped: bool

< Stats collection stopped due to saturation.