esp_idf_sys
Struct protocomm_security

#[repr(C)]
pub struct protocomm_security {
    pub ver: c_int,
    pub init: Option<unsafe extern "C" fn(handle: *mut protocomm_security_handle_t) -> esp_err_t>,
    pub cleanup: Option<unsafe extern "C" fn(handle: protocomm_security_handle_t) -> esp_err_t>,
    pub new_transport_session: Option<unsafe extern "C" fn(handle: protocomm_security_handle_t, session_id: u32) -> esp_err_t>,
    pub close_transport_session: Option<unsafe extern "C" fn(handle: protocomm_security_handle_t, session_id: u32) -> esp_err_t>,
    pub security_req_handler: Option<unsafe extern "C" fn(handle: protocomm_security_handle_t, sec_params: *const c_void, session_id: u32, inbuf: *const u8, inlen: isize, outbuf: *mut *mut u8, outlen: *mut isize, priv_data: *mut c_void) -> esp_err_t>,
    pub encrypt: Option<unsafe extern "C" fn(handle: protocomm_security_handle_t, session_id: u32, inbuf: *const u8, inlen: isize, outbuf: *mut *mut u8, outlen: *mut isize) -> esp_err_t>,
    pub decrypt: Option<unsafe extern "C" fn(handle: protocomm_security_handle_t, session_id: u32, inbuf: *const u8, inlen: isize, outbuf: *mut *mut u8, outlen: *mut isize) -> esp_err_t>,
}

@brief Protocomm security object structure.

The member functions are used for implementing secure protocomm sessions.

@note This structure should not have any dynamic members to allow re-entrancy

Fields
ver: c_int

Unique version number of security implementation

init: Option<unsafe extern "C" fn(handle: *mut protocomm_security_handle_t) -> esp_err_t>

Function for initializing/allocating security infrastructure

cleanup: Option<unsafe extern "C" fn(handle: protocomm_security_handle_t) -> esp_err_t>

Function for deallocating security infrastructure

new_transport_session: Option<unsafe extern "C" fn(handle: protocomm_security_handle_t, session_id: u32) -> esp_err_t>

Starts new secure transport session with specified ID

close_transport_session: Option<unsafe extern "C" fn(handle: protocomm_security_handle_t, session_id: u32) -> esp_err_t>

Closes a secure transport session with specified ID

security_req_handler: Option<unsafe extern "C" fn(handle: protocomm_security_handle_t, sec_params: *const c_void, session_id: u32, inbuf: *const u8, inlen: isize, outbuf: *mut *mut u8, outlen: *mut isize, priv_data: *mut c_void) -> esp_err_t>

Handler function for authenticating connection request and establishing secure session

encrypt: Option<unsafe extern "C" fn(handle: protocomm_security_handle_t, session_id: u32, inbuf: *const u8, inlen: isize, outbuf: *mut *mut u8, outlen: *mut isize) -> esp_err_t>

Function which implements the encryption algorithm

decrypt: Option<unsafe extern "C" fn(handle: protocomm_security_handle_t, session_id: u32, inbuf: *const u8, inlen: isize, outbuf: *mut *mut u8, outlen: *mut isize) -> esp_err_t>

Function which implements the decryption algorithm