esp_idf_sys
Struct wifi_init_config_t

#[repr(C)]
pub struct wifi_init_config_t {

    pub osi_funcs: *mut wifi_osi_funcs_t,
    pub wpa_crypto_funcs: wpa_crypto_funcs_t,
    pub static_rx_buf_num: c_int,
    pub dynamic_rx_buf_num: c_int,
    pub tx_buf_type: c_int,
    pub static_tx_buf_num: c_int,
    pub dynamic_tx_buf_num: c_int,
    pub rx_mgmt_buf_type: c_int,
    pub rx_mgmt_buf_num: c_int,
    pub cache_tx_buf_num: c_int,
    pub csi_enable: c_int,
    pub ampdu_rx_enable: c_int,
    pub ampdu_tx_enable: c_int,
    pub amsdu_tx_enable: c_int,
    pub nvs_enable: c_int,
    pub nano_enable: c_int,
    pub rx_ba_win: c_int,
    pub wifi_task_core_id: c_int,
    pub beacon_max_len: c_int,
    pub mgmt_sbuf_num: c_int,
    pub feature_caps: u64,
    pub sta_disconnected_pm: bool,
    pub espnow_max_encrypt_num: c_int,
    pub tx_hetb_queue_num: c_int,
    pub dump_hesigb_enable: bool,
    pub magic: c_int,

}

@brief WiFi stack configuration parameters passed to esp_wifi_init call.

Fields
osi_funcs: *mut wifi_osi_funcs_t

< WiFi OS functions

wpa_crypto_funcs: wpa_crypto_funcs_t

< WiFi station crypto functions when connect

static_rx_buf_num: c_int

< WiFi static RX buffer number

dynamic_rx_buf_num: c_int

< WiFi dynamic RX buffer number

tx_buf_type: c_int

< WiFi TX buffer type

static_tx_buf_num: c_int

< WiFi static TX buffer number

dynamic_tx_buf_num: c_int

< WiFi dynamic TX buffer number

rx_mgmt_buf_type: c_int

< WiFi RX MGMT buffer type

rx_mgmt_buf_num: c_int

< WiFi RX MGMT buffer number

cache_tx_buf_num: c_int

< WiFi TX cache buffer number

csi_enable: c_int

< WiFi channel state information enable flag

ampdu_rx_enable: c_int

< WiFi AMPDU RX feature enable flag

ampdu_tx_enable: c_int

< WiFi AMPDU TX feature enable flag

amsdu_tx_enable: c_int

< WiFi AMSDU TX feature enable flag

nvs_enable: c_int

< WiFi NVS flash enable flag

nano_enable: c_int

< Nano option for printf/scan family enable flag

rx_ba_win: c_int

< WiFi Block Ack RX window size

wifi_task_core_id: c_int

< WiFi Task Core ID

beacon_max_len: c_int

< WiFi softAP maximum length of the beacon

mgmt_sbuf_num: c_int

< WiFi management short buffer number, the minimum value is 6, the maximum value is 32

feature_caps: u64

< Enables additional WiFi features and capabilities

sta_disconnected_pm: bool

< WiFi Power Management for station at disconnected status

espnow_max_encrypt_num: c_int

< Maximum encrypt number of peers supported by espnow

tx_hetb_queue_num: c_int

< WiFi TX HE TB QUEUE number for STA HE TB PPDU transmission

dump_hesigb_enable: bool

< enable dump sigb field

magic: c_int

< WiFi init magic number, it should be the last field