esp_idf_sys

Struct wifi_nan_subscribe_cfg_t

#[repr(C)]
pub struct wifi_nan_subscribe_cfg_t {
    pub service_name: [c_char; 256],
    pub type_: wifi_nan_service_type_t,
    pub matching_filter: [c_char; 256],
    pub svc_info: [c_char; 64],
    pub _bitfield_align_1: [u8; 0],
    pub _bitfield_1: __BindgenBitfieldUnit<[u8; 1]>,
    pub ssi_len: u16,
    pub ssi: *mut u8,
}

@brief NAN Subscribe service configuration parameters

Fields
service_name: [c_char; 256]

< Service name identifier

type_: wifi_nan_service_type_t

< Service type

matching_filter: [c_char; 256]

< Comma separated filters for filtering services

svc_info: [c_char; 64]

< To be deprecated in next major release, use ssi instead

_bitfield_align_1: [u8; 0]_bitfield_1: __BindgenBitfieldUnit<[u8; 1]>ssi_len: u16

< Length of service specific info, maximum allowed length - ESP_WIFI_MAX_SVC_SSI_LEN

ssi: *mut u8

< Service Specific Info of type wifi_nan_wfa_ssi_t for WFA defined protocols, otherwise proprietary and defined by Applications

Implementations

impl wifi_nan_subscribe_cfg_t

pub fn single_match_event(&self) -> u8

pub fn set_single_match_event(&mut self, val: u8)

pub unsafe fn single_match_event_raw(this: *const Self) -> u8

pub unsafe fn set_single_match_event_raw(this: *mut Self, val: u8)

pub fn datapath_reqd(&self) -> u8

pub fn set_datapath_reqd(&mut self, val: u8)

pub unsafe fn datapath_reqd_raw(this: *const Self) -> u8

pub unsafe fn set_datapath_reqd_raw(this: *mut Self, val: u8)

pub fn fsd_reqd(&self) -> u8

pub fn set_fsd_reqd(&mut self, val: u8)

pub unsafe fn fsd_reqd_raw(this: *const Self) -> u8

pub unsafe fn set_fsd_reqd_raw(this: *mut Self, val: u8)

pub fn fsd_gas(&self) -> u8

pub fn set_fsd_gas(&mut self, val: u8)

pub unsafe fn fsd_gas_raw(this: *const Self) -> u8

pub unsafe fn set_fsd_gas_raw(this: *mut Self, val: u8)

pub fn reserved(&self) -> u8

pub fn set_reserved(&mut self, val: u8)

pub unsafe fn reserved_raw(this: *const Self) -> u8

pub unsafe fn set_reserved_raw(this: *mut Self, val: u8)

pub fn new_bitfield_1( single_match_event: u8, datapath_reqd: u8, fsd_reqd: u8, fsd_gas: u8, reserved: u8, ) -> __BindgenBitfieldUnit<[u8; 1]>