esp_idf_sys

Type Alias StaticRingbuffer_t

pub type StaticRingbuffer_t = xSTATIC_RINGBUFFER;

@brief Struct that is equivalent in size to the ring buffer’s data structure

The contents of this struct are not meant to be used directly. This structure is meant to be used when creating a statically allocated ring buffer where this struct is of the exact size required to store a ring buffer’s control data structure.

Aliased Type

#[repr(C)]
pub struct StaticRingbuffer_t {
    pub xDummy1: [usize; 2],
    pub uxDummy2: u32,
    pub pvDummy3: [*mut c_void; 11],
    pub xDummy4: i32,
    pub xDummy5: [xSTATIC_LIST; 2],
    pub pvDummy6: *mut c_void,
    pub muxDummy: spinlock_t,
}

Fields
xDummy1: [usize; 2]

@cond

uxDummy2: u32pvDummy3: [*mut c_void; 11]xDummy4: i32xDummy5: [xSTATIC_LIST; 2]pvDummy6: *mut c_voidmuxDummy: spinlock_t