esp_idf_sys
Type Alias esp_tls_cfg_server_t

pub type esp_tls_cfg_server_t = esp_tls_cfg_server;

@brief ESP-TLS Server configuration parameters

Aliased Type

struct esp_tls_cfg_server_t {

    pub alpn_protos: *mut *const u8,
    pub __bindgen_anon_1: esp_tls_cfg_server__bindgen_ty_1,
    pub __bindgen_anon_2: esp_tls_cfg_server__bindgen_ty_2,
    pub __bindgen_anon_3: esp_tls_cfg_server__bindgen_ty_3,
    pub __bindgen_anon_4: esp_tls_cfg_server__bindgen_ty_4,
    pub __bindgen_anon_5: esp_tls_cfg_server__bindgen_ty_5,
    pub __bindgen_anon_6: esp_tls_cfg_server__bindgen_ty_6,
    pub serverkey_password: *const u8,
    pub serverkey_password_len: u32,
    pub use_ecdsa_peripheral: bool,
    pub ecdsa_key_efuse_blk: u8,
    pub use_secure_element: bool,
    pub userdata: *mut c_void,

}

Fields
alpn_protos: *mut *const u8

< Application protocols required for HTTP2. If HTTP2/ALPN support is required, a list of protocols that should be negotiated. The format is length followed by protocol name. For the most common cases the following is ok: const char **alpn_protos = { “h2”, NULL };

where ‘h2’ is the protocol name

__bindgen_anon_1: esp_tls_cfg_server__bindgen_ty_1__bindgen_anon_2: esp_tls_cfg_server__bindgen_ty_2__bindgen_anon_3: esp_tls_cfg_server__bindgen_ty_3__bindgen_anon_4: esp_tls_cfg_server__bindgen_ty_4__bindgen_anon_5: esp_tls_cfg_server__bindgen_ty_5__bindgen_anon_6: esp_tls_cfg_server__bindgen_ty_6serverkey_password: *const u8

< Server key decryption password string

serverkey_password_len: u32

< String length of the password pointed to by serverkey_password

use_ecdsa_peripheral: bool

< Use ECDSA peripheral to use private key

ecdsa_key_efuse_blk: u8

< The efuse block where ECDSA key is stored

use_secure_element: bool

< Enable this option to use secure element or atecc608a chip

userdata: *mut c_void

< User data to be added to the ssl context. Can be retrieved by callbacks