Type Alias esp_idf_sys::esp_tls_t

pub type esp_tls_t = esp_tls;

@brief ESP-TLS Connection Handle

Aliased Type

struct esp_tls_t {

    pub ssl: mbedtls_ssl_context,
    pub entropy: mbedtls_entropy_context,
    pub ctr_drbg: mbedtls_ctr_drbg_context,
    pub conf: mbedtls_ssl_config,
    pub server_fd: mbedtls_net_context,
    pub cacert: mbedtls_x509_crt,
    pub cacert_ptr: *mut mbedtls_x509_crt,
    pub clientcert: mbedtls_x509_crt,
    pub clientkey: mbedtls_pk_context,
    pub sockfd: i32,
    pub read: Option<unsafe extern "C" fn(_: *mut esp_tls, _: *mut i8, _: usize) -> isize>,
    pub write: Option<unsafe extern "C" fn(_: *mut esp_tls, _: *const i8, _: usize) -> isize>,
    pub conn_state: u32,
    pub rset: _types_fd_set,
    pub wset: _types_fd_set,
    pub is_tls: bool,
    pub role: u32,
    pub error_handle: *mut esp_tls_last_error,

}

Fields
ssl: mbedtls_ssl_context

< TLS/SSL context

entropy: mbedtls_entropy_context

< mbedTLS entropy context structure

ctr_drbg: mbedtls_ctr_drbg_context

< mbedTLS ctr drbg context structure. CTR_DRBG is deterministic random bit generation based on AES-256

conf: mbedtls_ssl_config

< TLS/SSL configuration to be shared between mbedtls_ssl_context structures

server_fd: mbedtls_net_context

< mbedTLS wrapper type for sockets

cacert: mbedtls_x509_crt

< Container for the X.509 CA certificate

cacert_ptr: *mut mbedtls_x509_crt

< Pointer to the cacert being used.

clientcert: mbedtls_x509_crt

< Container for the X.509 client certificate

clientkey: mbedtls_pk_context

< Container for the private key of the client certificate

sockfd: i32

< Underlying socket file descriptor.

read: Option<unsafe extern "C" fn(_: *mut esp_tls, _: *mut i8, _: usize) -> isize>

< Callback function for reading data from TLS/SSL connection.

write: Option<unsafe extern "C" fn(_: *mut esp_tls, _: *const i8, _: usize) -> isize>

< Callback function for writing data to TLS/SSL connection.

conn_state: u32

< ESP-TLS Connection state

rset: _types_fd_set

< read file descriptors

wset: _types_fd_set

< write file descriptors

is_tls: bool

< indicates connection type (TLS or NON-TLS)

role: u32

< esp-tls role

ESP_TLS_CLIENT

ESP_TLS_SERVER

error_handle: *mut esp_tls_last_error

< handle to error descriptor