Type Definition esp_idf_sys::mbedtls_ssl_set_timer_t

pub type mbedtls_ssl_set_timer_t = Option<unsafe extern "C" fn(ctx: *mut c_void, int_ms: u32, fin_ms: u32)>;

\brief Callback type: set a pair of timers/delays to watch

\param ctx Context pointer \param int_ms Intermediate delay in milliseconds \param fin_ms Final delay in milliseconds 0 cancels the current timer.

\note This callback must at least store the necessary information for the associated \c mbedtls_ssl_get_timer_t callback to return correct information.

\note If using a event-driven style of programming, an event must be generated when the final delay is passed. The event must cause a call to \c mbedtls_ssl_handshake() with the proper SSL context to be scheduled. Care must be taken to ensure that at most one such call happens at a time.

\note Only one timer at a time must be running. Calling this function while a timer is running must cancel it. Cancelled timers must not generate any event.