esp_idf_sys
Type Alias psa_algorithm_t

pub type psa_algorithm_t = u32;

\brief Encoding of a cryptographic algorithm.

Values of this type are generally constructed by macros called PSA_ALG_xxx.

For algorithms that can be applied to multiple key types, this type does not encode the key type. For example, for symmetric ciphers based on a block cipher, #psa_algorithm_t encodes the block cipher mode and the padding mode while the block cipher itself is encoded via #psa_key_type_t.

\note Values of this type are encoded in the persistent key store. Any changes to existing values will require bumping the storage format version and providing a translation when reading the old format.