esp_idf_sys

Type Alias psa_key_derivation_operation_t

pub type psa_key_derivation_operation_t = psa_key_derivation_s;

The type of the state data structure for key derivation operations.

Before calling any function on a key derivation operation object, the application must initialize it by any of the following means:

Set the structure to all-bits-zero, for example: \code psa_key_derivation_operation_t operation; memset(&operation, 0, sizeof(operation)); \endcode

Initialize the structure to logical zero values, for example: \code psa_key_derivation_operation_t operation = {0}; \endcode

Initialize the structure to the initializer #PSA_KEY_DERIVATION_OPERATION_INIT, for example: \code psa_key_derivation_operation_t operation = PSA_KEY_DERIVATION_OPERATION_INIT; \endcode

Assign the result of the function psa_key_derivation_operation_init() to the structure, for example: \code psa_key_derivation_operation_t operation; operation = psa_key_derivation_operation_init(); \endcode

This is an implementation-defined \c struct. Applications should not make any assumptions about the content of this structure. Implementation details can change in future versions without notice.

Aliased Type

#[repr(C)]
pub struct psa_key_derivation_operation_t {
    pub private_alg: u32,
    pub _bitfield_align_1: [u8; 0],
    pub _bitfield_1: __BindgenBitfieldUnit<[u8; 1]>,
    pub private_capacity: usize,
    pub private_ctx: psa_driver_key_derivation_context_t,
}

Fields
private_alg: u32_bitfield_align_1: [u8; 0]_bitfield_1: __BindgenBitfieldUnit<[u8; 1]>private_capacity: usizeprivate_ctx: psa_driver_key_derivation_context_t