esp_idf_sys
Type Alias wifi_prov_scheme_t

pub type wifi_prov_scheme_t = wifi_prov_scheme;

@brief Structure for specifying the provisioning scheme to be followed by the manager

@note Ready to use schemes are available: - wifi_prov_scheme_ble : for provisioning over BLE transport + GATT server - wifi_prov_scheme_softap : for provisioning over SoftAP transport + HTTP server - wifi_prov_scheme_console : for provisioning over Serial UART transport + Console (for debugging)

Aliased Type

struct wifi_prov_scheme_t {
    pub prov_start: Option<unsafe extern "C" fn(_: *mut protocomm, _: *mut c_void) -> i32>,
    pub prov_stop: Option<unsafe extern "C" fn(_: *mut protocomm) -> i32>,
    pub new_config: Option<unsafe extern "C" fn() -> *mut c_void>,
    pub delete_config: Option<unsafe extern "C" fn(_: *mut c_void)>,
    pub set_config_service: Option<unsafe extern "C" fn(_: *mut c_void, _: *const u8, _: *const u8) -> i32>,
    pub set_config_endpoint: Option<unsafe extern "C" fn(_: *mut c_void, _: *const u8, _: u16) -> i32>,
    pub wifi_mode: u32,
}

Fields
prov_start: Option<unsafe extern "C" fn(_: *mut protocomm, _: *mut c_void) -> i32>

Function which is to be called by the manager when it is to start the provisioning service associated with a protocomm instance and a scheme specific configuration

prov_stop: Option<unsafe extern "C" fn(_: *mut protocomm) -> i32>

Function which is to be called by the manager to stop the provisioning service previously associated with a protocomm instance

new_config: Option<unsafe extern "C" fn() -> *mut c_void>

Function which is to be called by the manager to generate a new configuration for the provisioning service, that is to be passed to prov_start()

delete_config: Option<unsafe extern "C" fn(_: *mut c_void)>

Function which is to be called by the manager to delete a configuration generated using new_config()

set_config_service: Option<unsafe extern "C" fn(_: *mut c_void, _: *const u8, _: *const u8) -> i32>

Function which is to be called by the manager to set the service name and key values in the configuration structure

set_config_endpoint: Option<unsafe extern "C" fn(_: *mut c_void, _: *const u8, _: u16) -> i32>

Function which is to be called by the manager to set a protocomm endpoint with an identifying name and UUID in the configuration structure

wifi_mode: u32

Sets mode of operation of Wi-Fi during provisioning This is set to :

WIFI_MODE_APSTA for SoftAP transport

WIFI_MODE_STA for BLE transport