esp_idf_sys
Union mesh_event_info_t

#[repr(C)]
pub union mesh_event_info_t {

    pub channel_switch: mesh_event_channel_switch_t,
    pub child_connected: mesh_event_child_connected_t,
    pub child_disconnected: mesh_event_child_disconnected_t,
    pub routing_table: mesh_event_routing_table_change_t,
    pub connected: mesh_event_connected_t,
    pub disconnected: mesh_event_disconnected_t,
    pub no_parent: mesh_event_no_parent_found_t,
    pub layer_change: mesh_event_layer_change_t,
    pub toDS_state: mesh_event_toDS_state_t,
    pub vote_started: mesh_event_vote_started_t,
    pub root_addr: mesh_event_root_address_t,
    pub switch_req: mesh_event_root_switch_req_t,
    pub root_conflict: mesh_event_root_conflict_t,
    pub root_fixed: mesh_event_root_fixed_t,
    pub scan_done: mesh_event_scan_done_t,
    pub network_state: mesh_event_network_state_t,
    pub find_network: mesh_event_find_network_t,
    pub router_switch: mesh_event_router_switch_t,
    pub ps_duty: mesh_event_ps_duty_t,

}

@brief Mesh event information

Fields
channel_switch: mesh_event_channel_switch_t

< channel switch

child_connected: mesh_event_child_connected_t

< child connected

child_disconnected: mesh_event_child_disconnected_t

< child disconnected

routing_table: mesh_event_routing_table_change_t

< routing table change

connected: mesh_event_connected_t

< parent connected

disconnected: mesh_event_disconnected_t

< parent disconnected

no_parent: mesh_event_no_parent_found_t

< no parent found

layer_change: mesh_event_layer_change_t

< layer change

toDS_state: mesh_event_toDS_state_t

< toDS state, devices shall check this state firstly before trying to send packets to external IP network. This state indicates right now whether the root is capable of sending packets out. If not, devices had better to wait until this state changes to be MESH_TODS_REACHABLE.

vote_started: mesh_event_vote_started_t

< vote started

root_addr: mesh_event_root_address_t

< root address

switch_req: mesh_event_root_switch_req_t

< root switch request

root_conflict: mesh_event_root_conflict_t

< other powerful root

root_fixed: mesh_event_root_fixed_t

< fixed root

scan_done: mesh_event_scan_done_t

< scan done

network_state: mesh_event_network_state_t

< network state, such as whether current mesh network has a root.

find_network: mesh_event_find_network_t

< network found that can join

router_switch: mesh_event_router_switch_t

< new router information

ps_duty: mesh_event_ps_duty_t

< PS duty information