Function esp_idf_sys::esp_ota_get_next_update_partition

pub unsafe extern "C" fn esp_ota_get_next_update_partition(
    start_from: *const esp_partition_t
) -> *const esp_partition_t

@brief Return the next OTA app partition which should be written with a new firmware.

Call this function to find an OTA app partition which can be passed to esp_ota_begin().

Finds next partition round-robin, starting from the current running partition.

@param start_from If set, treat this partition info as describing the current running partition. Can be NULL, in which case esp_ota_get_running_partition() is used to find the currently running partition. The result of this function is never the same as this argument.

@return Pointer to info for partition which should be updated next. NULL result indicates invalid OTA data partition, or that no eligible OTA app slot partition was found.