Function esp_idf_sys::mbedtls_ecdsa_write_signature_det

pub unsafe extern "C" fn mbedtls_ecdsa_write_signature_det(
    ctx: *mut mbedtls_ecdsa_context,
    hash: *const c_uchar,
    hlen: usize,
    sig: *mut c_uchar,
    slen: *mut usize,
    md_alg: mbedtls_md_type_t
) -> c_int

\brief This function computes an ECDSA signature and writes it to a buffer, serialized as defined in RFC-4492: Elliptic Curve Cryptography (ECC) Cipher Suites for Transport Layer Security (TLS).

             The deterministic version is defined in <em>RFC-6979:
             Deterministic Usage of the Digital Signature Algorithm (DSA)
             and Elliptic Curve Digital Signature Algorithm (ECDSA)</em>.

\warning It is not thread-safe to use the same context in multiple threads.

\note If the bitlength of the message hash is larger than the bitlength of the group order, then the hash is truncated as defined in Standards for Efficient Cryptography Group (SECG): SEC1 Elliptic Curve Cryptography, section 4.1.3, step 5.

\see ecp.h

\deprecated Superseded by mbedtls_ecdsa_write_signature() in Mbed TLS version 2.0 and later.

\param ctx The ECDSA context to use. This must be initialized and have a group and private key bound to it, for example via mbedtls_ecdsa_genkey() or mbedtls_ecdsa_from_keypair(). \param hash The message hash to be signed. This must be a readable buffer of length \p blen Bytes. \param hlen The length of the hash \p hash in Bytes. \param sig The buffer to which to write the signature. This must be a writable buffer of length at least twice as large as the size of the curve used, plus 9. For example, 73 Bytes if a 256-bit curve is used. A buffer length of #MBEDTLS_ECDSA_MAX_LEN is always safe. \param slen The address at which to store the actual length of the signature written. Must not be \c NULL. \param md_alg The message digest that was used to hash the message.

\return \c 0 on success. \return An \c MBEDTLS_ERR_ECP_XXX, \c MBEDTLS_ERR_MPI_XXX or \c MBEDTLS_ERR_ASN1_XXX error code on failure.