Function esp_idf_sys::mbedtls_rsa_pkcs1_encrypt

pub unsafe extern "C" fn mbedtls_rsa_pkcs1_encrypt(
    ctx: *mut mbedtls_rsa_context, 
    f_rng: Option<unsafe extern "C" fn(arg1: *mut c_void, arg2: *mut c_uchar, arg3: size_t) -> c_int>, 
    p_rng: *mut c_void, 
    mode: c_int, 
    ilen: size_t, 
    input: *const c_uchar, 
    output: *mut c_uchar
) -> c_int

\brief This function adds the message padding, then performs an RSA operation.

            It is the generic wrapper for performing a PKCS#1 encryption
            operation using the \p mode from the context.

\deprecated It is deprecated and discouraged to call this function in #MBEDTLS_RSA_PRIVATE mode. Future versions of the library are likely to remove the \p mode argument and have it implicitly set to #MBEDTLS_RSA_PUBLIC.

\note Alternative implementations of RSA need not support mode being set to #MBEDTLS_RSA_PRIVATE and might instead return #MBEDTLS_ERR_PLATFORM_FEATURE_UNSUPPORTED.

\param ctx The initialized RSA context to use. \param f_rng The RNG to use. It is mandatory for PKCS#1 v2.1 padding encoding, and for PKCS#1 v1.5 padding encoding when used with \p mode set to #MBEDTLS_RSA_PUBLIC. For PKCS#1 v1.5 padding encoding and \p mode set to #MBEDTLS_RSA_PRIVATE, it is used for blinding and should be provided in this case; see mbedtls_rsa_private() for more. \param p_rng The RNG context to be passed to \p f_rng. May be \c NULL if \p f_rng is \c NULL or if \p f_rng doesn’t need a context argument. \param mode The mode of operation. This must be either #MBEDTLS_RSA_PUBLIC or #MBEDTLS_RSA_PRIVATE (deprecated). \param ilen The length of the plaintext in Bytes. \param input The input data to encrypt. This must be a readable buffer of size \p ilen Bytes. This must not be \c NULL. \param output The output buffer. This must be a writable buffer of length \c ctx->len Bytes. For example, \c 256 Bytes for an 2048-bit RSA modulus.

\return \c 0 on success. \return An \c MBEDTLS_ERR_RSA_XXX error code on failure.