Function esp_idf_sys::mbedtls_ssl_conf_curves

pub unsafe extern "C" fn mbedtls_ssl_conf_curves(
    conf: *mut mbedtls_ssl_config, 
    curves: *const mbedtls_ecp_group_id
)

\brief Set the allowed curves in order of preference. (Default: all defined curves in order of decreasing size.)

            On server: this only affects selection of the ECDHE curve;
            the curves used for ECDH and ECDSA are determined by the
            list of available certificates instead.

            On client: this affects the list of curves offered for any
            use. The server can override our preference order.

            Both sides: limits the set of curves accepted for use in
            ECDHE and in the peer's end-entity certificate.

\note This has no influence on which curves are allowed inside the certificate chains, see \c mbedtls_ssl_conf_cert_profile() for that. For the end-entity certificate however, the key will be accepted only if it is allowed both by this list and by the cert profile.

\note This list should be ordered by decreasing preference (preferred curve first).

\param conf SSL configuration \param curves Ordered list of allowed curves, terminated by MBEDTLS_ECP_DP_NONE.