esp_idf_sys
Function mbedtls_x509_crt_verify_restartable

pub unsafe extern "C" fn mbedtls_x509_crt_verify_restartable(
    crt: *mut mbedtls_x509_crt,
    trust_ca: *mut mbedtls_x509_crt,
    ca_crl: *mut mbedtls_x509_crl,
    profile: *const mbedtls_x509_crt_profile,
    cn: *const c_char,
    flags: *mut u32,
    f_vrfy: Option<unsafe extern "C" fn(arg1: *mut c_void, arg2: *mut mbedtls_x509_crt, arg3: c_int, arg4: *mut u32) -> c_int>,
    p_vrfy: *mut c_void,
    rs_ctx: *mut mbedtls_x509_crt_restart_ctx,
) -> c_int

\brief Restartable version of \c mbedtls_crt_verify_with_profile()

\note Performs the same job as \c mbedtls_crt_verify_with_profile() but can return early and restart according to the limit set with \c mbedtls_ecp_set_max_ops() to reduce blocking.

\param crt The certificate chain to be verified. \param trust_ca The list of trusted CAs. \param ca_crl The list of CRLs for trusted CAs. \param profile The security profile to use for the verification. \param cn The expected Common Name. This may be \c NULL if the CN need not be verified. \param flags The address at which to store the result of the verification. If the verification couldn’t be completed, the flag value is set to (uint32_t) -1. \param f_vrfy The verification callback to use. See the documentation of mbedtls_x509_crt_verify() for more information. \param p_vrfy The context to be passed to \p f_vrfy. \param rs_ctx The restart context to use. This may be set to \c NULL to disable restartable ECC.

\return See \c mbedtls_crt_verify_with_profile(), or \return #MBEDTLS_ERR_ECP_IN_PROGRESS if maximum number of operations was reached: see \c mbedtls_ecp_set_max_ops().