Function esp_idf_sys::nvs_entry_find

pub unsafe extern "C" fn nvs_entry_find(
    part_name: *const c_char, 
    namespace_name: *const c_char, 
    type_: nvs_type_t
) -> nvs_iterator_t

@brief Create an iterator to enumerate NVS entries based on one or more parameters

\code{c} // Example of listing all the key-value pairs of any type under specified partition and namespace nvs_iterator_t it = nvs_entry_find(partition, namespace, NVS_TYPE_ANY); while (it != NULL) { nvs_entry_info_t info; nvs_entry_info(it, &info); it = nvs_entry_next(it); printf(“key ‘%s’, type ‘%d’ \n”, info.key, info.type); }; // Note: no need to release iterator obtained from nvs_entry_find function when // nvs_entry_find or nvs_entry_next function return NULL, indicating no other // element for specified criteria was found. } \endcode

@param[in] part_name Partition name

@param[in] namespace_name Set this value if looking for entries with a specific namespace. Pass NULL otherwise.

@param[in] type One of nvs_type_t values.

@return Iterator used to enumerate all the entries found, or NULL if no entry satisfying criteria was found. Iterator obtained through this function has to be released using nvs_release_iterator when not used any more.