esp_idf_sys
Function psa_reset_key_attributes

pub unsafe extern "C" fn psa_reset_key_attributes(
    attributes: *mut psa_key_attributes_t,
)

Reset a key attribute structure to a freshly initialized state.

You must initialize the attribute structure as described in the documentation of the type #psa_key_attributes_t before calling this function. Once the structure has been initialized, you may call this function at any time.

This function frees any auxiliary resources that the structure may contain.

\param[in,out] attributes The attribute structure to reset.