Function esp_idf_sys::vTaskDelete

pub unsafe extern "C" fn vTaskDelete(xTaskToDelete: TaskHandle_t)

Remove a task from the RTOS real time kernel’s management. The task being deleted will be removed from all ready, blocked, suspended and event lists.

INCLUDE_vTaskDelete must be defined as 1 for this function to be available. See the configuration section for more information.

NOTE: The idle task is responsible for freeing the kernel allocated memory from tasks that have been deleted. It is therefore important that the idle task is not starved of microcontroller processing time if your application makes any calls to vTaskDelete (). Memory allocated by the task code is not automatically freed, and should be freed before the task is deleted.

See the demo application file death.c for sample code that utilises vTaskDelete ().

@param xTaskToDelete The handle of the task to be deleted. Passing NULL will cause the calling task to be deleted.

Example usage: @code{c} void vOtherFunction( void ) { TaskHandle_t xHandle;

 // Create the task, storing the handle.
 xTaskCreate( vTaskCode, "NAME", STACK_SIZE, NULL, tskIDLE_PRIORITY, &xHandle );

 // Use the handle to delete the task.
 vTaskDelete( xHandle );

} @endcode \ingroup Tasks