esp_idf_sys
Function wifi_prov_mgr_configure_sta

pub unsafe extern "C" fn wifi_prov_mgr_configure_sta(
    wifi_cfg: *mut wifi_config_t,
) -> esp_err_t

@brief Runs Wi-Fi as Station with the supplied configuration

Configures the Wi-Fi station mode to connect to the AP with SSID and password specified in config structure and sets Wi-Fi to run as station.

This is automatically called by provisioning service upon receiving new credentials.

If credentials are to be supplied to the manager via a different mode other than through protocomm, then this API needs to be called.

Event WIFI_PROV_CRED_RECV is emitted after credentials have been applied and Wi-Fi station started

@param[in] wifi_cfg Pointer to Wi-Fi configuration structure

@return

ESP_OK : Wi-Fi configured and started successfully

ESP_FAIL : Failed to set configuration