Struct esp_idf_sys::wifi_pkt_rx_ctrl_t

#[repr(C)]

#[repr(align(4))]
pub struct wifi_pkt_rx_ctrl_t {
    pub _bitfield_align_1: [u32; 0],
    pub _bitfield_1: __BindgenBitfieldUnit<[u8; 48]>,
}

@brief Received packet radio metadata header, this is the common header at the beginning of all promiscuous mode RX callback buffers

Fields
_bitfield_align_1: [u32; 0]_bitfield_1: __BindgenBitfieldUnit<[u8; 48]>
Implementations

impl wifi_pkt_rx_ctrl_t

pub fn rssi(&self) -> c_int

pub fn set_rssi(&mut self, val: c_int)

pub fn rate(&self) -> c_uint

pub fn set_rate(&mut self, val: c_uint)

pub fn sig_mode(&self) -> c_uint

pub fn set_sig_mode(&mut self, val: c_uint)

pub fn mcs(&self) -> c_uint

pub fn set_mcs(&mut self, val: c_uint)

pub fn cwb(&self) -> c_uint

pub fn set_cwb(&mut self, val: c_uint)

pub fn smoothing(&self) -> c_uint

pub fn set_smoothing(&mut self, val: c_uint)

pub fn not_sounding(&self) -> c_uint

pub fn set_not_sounding(&mut self, val: c_uint)

pub fn aggregation(&self) -> c_uint

pub fn set_aggregation(&mut self, val: c_uint)

pub fn stbc(&self) -> c_uint

pub fn set_stbc(&mut self, val: c_uint)

pub fn fec_coding(&self) -> c_uint

pub fn set_fec_coding(&mut self, val: c_uint)

pub fn sgi(&self) -> c_uint

pub fn set_sgi(&mut self, val: c_uint)

pub fn ampdu_cnt(&self) -> c_uint

pub fn set_ampdu_cnt(&mut self, val: c_uint)

pub fn channel(&self) -> c_uint

pub fn set_channel(&mut self, val: c_uint)

pub fn secondary_channel(&self) -> c_uint

pub fn set_secondary_channel(&mut self, val: c_uint)

pub fn timestamp(&self) -> c_uint

pub fn set_timestamp(&mut self, val: c_uint)

pub fn noise_floor(&self) -> c_int

pub fn set_noise_floor(&mut self, val: c_int)

pub fn ant(&self) -> c_uint

pub fn set_ant(&mut self, val: c_uint)

pub fn sig_len(&self) -> c_uint

pub fn set_sig_len(&mut self, val: c_uint)

pub fn rx_state(&self) -> c_uint

pub fn set_rx_state(&mut self, val: c_uint)