Struct esp_idf_sys::wifi_sta_config_t

#[repr(C)]
pub struct wifi_sta_config_t {

    pub ssid: [u8; 32],
    pub password: [u8; 64],
    pub scan_method: wifi_scan_method_t,
    pub bssid_set: bool,
    pub bssid: [u8; 6],
    pub channel: u8,
    pub listen_interval: u16,
    pub sort_method: wifi_sort_method_t,
    pub threshold: wifi_scan_threshold_t,
    pub pmf_cfg: wifi_pmf_config_t,
    pub _bitfield_align_1: [u32; 0],
    pub _bitfield_1: __BindgenBitfieldUnit<[u8; 4]>,
    pub sae_pwe_h2e: wifi_sae_pwe_method_t,

}

@brief STA configuration settings for the ESP32

Fields
ssid: [u8; 32]

< SSID of target AP.

password: [u8; 64]

< Password of target AP.

scan_method: wifi_scan_method_t

< do all channel scan or fast scan

bssid_set: bool

< whether set MAC address of target AP or not. Generally, station_config.bssid_set needs to be 0; and it needs to be 1 only when users need to check the MAC address of the AP.

bssid: [u8; 6]

< MAC address of target AP

channel: u8

< channel of target AP. Set to 1~13 to scan starting from the specified channel before connecting to AP. If the channel of AP is unknown, set it to 0.

listen_interval: u16

< Listen interval for ESP32 station to receive beacon when WIFI_PS_MAX_MODEM is set. Units: AP beacon intervals. Defaults to 3 if set to 0.

sort_method: wifi_sort_method_t

< sort the connect AP in the list by rssi or security mode

threshold: wifi_scan_threshold_t

< When sort_method is set, only APs which have an auth mode that is more secure than the selected auth mode and a signal stronger than the minimum RSSI will be used.

pmf_cfg: wifi_pmf_config_t

< Configuration for Protected Management Frame. Will be advertized in RSN Capabilities in RSN IE.

_bitfield_align_1: [u32; 0]_bitfield_1: __BindgenBitfieldUnit<[u8; 4]>sae_pwe_h2e: wifi_sae_pwe_method_t

< Whether SAE hash to element is enabled

Implementations

impl wifi_sta_config_t

pub fn rm_enabled(&self) -> u32

pub fn set_rm_enabled(&mut self, val: u32)

pub fn btm_enabled(&self) -> u32

pub fn set_btm_enabled(&mut self, val: u32)

pub fn mbo_enabled(&self) -> u32

pub fn set_mbo_enabled(&mut self, val: u32)

pub fn reserved(&self) -> u32

pub fn set_reserved(&mut self, val: u32)

pub fn new_bitfield_1( rm_enabled: u32, btm_enabled: u32, mbo_enabled: u32, reserved: u32 ) -> __BindgenBitfieldUnit<[u8; 4]>