esp_idf_sys
Type Alias httpd_uri_t

pub type httpd_uri_t = httpd_uri;

@brief Structure for URI handler

Aliased Type

struct httpd_uri_t {
    pub uri: *const u8,
    pub method: u32,
    pub handler: Option<unsafe extern "C" fn(_: *mut httpd_req) -> i32>,
    pub user_ctx: *mut c_void,
}

Fields
uri: *const u8

< The URI to handle

method: u32

< Method supported by the URI, HTTP_ANY for wildcard method to support all methods

handler: Option<unsafe extern "C" fn(_: *mut httpd_req) -> i32>

Handler to call for supported request method. This must return ESP_OK, or else the underlying socket will be closed.

user_ctx: *mut c_void

Pointer to user context data which will be available to handler