Type Alias esp_idf_sys::mbedtls_psa_stats_t

pub type mbedtls_psa_stats_t = mbedtls_psa_stats_s;

\brief Statistics about resource consumption related to the PSA keystore.

\note The content of this structure is not part of the stable API and ABI of Mbed TLS and may change arbitrarily from version to version.

Aliased Type

struct mbedtls_psa_stats_t {
    pub private_volatile_slots: usize,
    pub private_persistent_slots: usize,
    pub private_external_slots: usize,
    pub private_half_filled_slots: usize,
    pub private_cache_slots: usize,
    pub private_empty_slots: usize,
    pub private_locked_slots: usize,
    pub private_max_open_internal_key_id: u32,
    pub private_max_open_external_key_id: u32,
}

Fields
private_volatile_slots: usizeprivate_persistent_slots: usizeprivate_external_slots: usizeprivate_half_filled_slots: usizeprivate_cache_slots: usizeprivate_empty_slots: usizeprivate_locked_slots: usizeprivate_max_open_internal_key_id: u32private_max_open_external_key_id: u32